% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archigos_doc.R
\docType{data}
\name{archigos}
\alias{archigos}
\title{The Archigos dataset of political regimes, v. 4.1}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3381 rows and 23 columns.}
\usage{
archigos
}
\description{
The ARCHIGOS dataset of political leaders, v. 4.1. Described in Henk E.
Goemans, Kristian Skrede Gleditsch, and Giacomo Chiozza. ".Introducing
Archigos: A Dataset of Political Leaders." Journal of Peace Research March 2009
vol. 46 no. 2 269-283. Original data and codebook available for download here:
\url{http://www.rochester.edu/college/faculty/hgoemans/data.htm}
}
\section{Variable descriptions}{


\describe{ \item{country_name}{Standardized country name. This is the same
across all datasets in this package, so you can always join them by
country_name and year. Character with 188 distinct values. Most common: France
(118), Japan (83), Switzerland (140). NAs = 0.}


\item{GWn}{Gleditsch-Ward numeric country code. See Gleditsch and Ward (1999).
Numeric. Max = 950, min = 2, distinct = 188, mean = 364.043, sd = 248.666, NAs
= 0.}


\item{leader}{Leader name. Character with 2722 distinct values. Most common:
Briand (6), del Castillo (6), Pats (6). NAs = 0.}


\item{obsid}{Observation ID. Factor with 3381 levels. Most common: AFG-1868
(1), AFG-1879-1 (1), AFG-1879-2 (1), AFG-1880 (1), AFG-1901 (1), AFG-1920 (1),
AFG-1929-1 (1), AFG-1929-2 (1), AFG-1929-3 (1), AFG-1933 (1), AFG-1946 (1),
AFG-1953 (1), AFG-1963 (1), AFG-1965 (1), AFG-1967-1 (1), AFG-1967-2 (1),
AFG-1971 (1), AFG-1972 (1), AFG-1973 (1), AFG-1978 (1), AFG-1979-1 (1),
AFG-1979-2 (1), AFG-1986 (1), AFG-1992-1 (1), AFG-1992-2 (1), AFG-1996 (1),
AFG-2001 (1), AFG-2014 (1), ALB-1912 (1), ALB-1914-1 (1), ALB-1914-2 (1),
ALB-1914-3 (1), ALB-1918 (1), ALB-1920-1 (1), ALB-1920-2 (1), ALB-1921-1 (1),
ALB-1921-2 (1), ALB-1921-3 (1), ALB-1921-4 (1), ALB-1922 (1), ALB-1924-1 (1),
ALB-1924-2 (1), ALB-1924-3 (1), ALB-1924-4 (1), ALB-1925 (1), ALB-1944 (1),
ALB-1985 (1), ALB-1992 (1), ALB-1997 (1), ALB-1998 (1), ALB-1999 (1),
ALB-2002-1 (1), ALB-2002-2 (1), ALB-2005 (1), ALB-2013 (1), ALG-1962-1 (1),
ALG-1962-2 (1), ALG-1965 (1), ALG-1978 (1), ALG-1979 (1), ALG-1992-1 (1),
ALG-1992-2 (1), ALG-1994 (1), ALG-1999 (1), ANG-1975 (1), ANG-1979 (1),
ARG-1874 (1), ARG-1880 (1), ARG-1886 (1), ARG-1890 (1), ARG-1892 (1), ARG-1895
(1), ARG-1898 (1), ARG-1904 (1), ARG-1906 (1), ARG-1910 (1), ARG-1913 (1),
ARG-1916 (1), ARG-1922 (1), ARG-1928 (1), ARG-1930 (1), ARG-1932 (1), ARG-1938
(1), ARG-1942 (1), ARG-1943-1 (1), ARG-1943-2 (1), ARG-1944 (1), ARG-1946 (1),
ARG-1955-1 (1), ARG-1955-2 (1), ARG-1958 (1), ARG-1962 (1), ARG-1963 (1),
ARG-1966 (1), ARG-1970-1 (1), ARG-1970-2 (1), ARG-1971 (1), ARG-1973-1 (1),
ARG-1973-2 (1), ARG-1973-3 (1), ARG-1974 (1), ARG-1976 (1), ARG-1981-1 (1),
ARG-1981-2 (1), ARG-1981-3 (1), ARG-1982-1 (1), ARG-1982-2 (1), ARG-1983 (1),
ARG-1989 (1), ARG-1999 (1), ARG-2001-1 (1), ARG-2001-2 (1), ARG-2002 (1),
ARG-2003 (1), ARG-2007 (1), ARM-1991 (1), ARM-1998 (1), ARM-2008 (1), AUH-1848
(1), AUH-1916 (1), AUL-1901 (1), AUL-1903 (1), AUL-1904-1 (1), AUL-1904-2 (1),
AUL-1905 (1), AUL-1908 (1), AUL-1909 (1), AUL-1910 (1), AUL-1913 (1), AUL-1914
(1), AUL-1915 (1), AUL-1923 (1), AUL-1929 (1), AUL-1932 (1), AUL-1939-1 (1),
AUL-1939-2 (1), AUL-1941-1 (1), AUL-1941-2 (1), AUL-1945-1 (1), AUL-1945-2 (1),
AUL-1949 (1), AUL-1966 (1), AUL-1967 (1), AUL-1968 (1), AUL-1971 (1), AUL-1972
(1), AUL-1975 (1), AUL-1983 (1), AUL-1991 (1), AUL-1996 (1), AUL-2007 (1),
AUL-2010 (1), AUL-2013-1 (1), AUL-2013-2 (1), AUS-1918 (1), AUS-1920 (1),
AUS-1921 (1), AUS-1922 (1), AUS-1924 (1), AUS-1926 (1), AUS-1929-1 (1),
AUS-1929-2 (1), AUS-1930-1 (1), AUS-1930-2 (1), AUS-1931 (1), AUS-1932 (1),
AUS-1934 (1), AUS-1938 (1), AUS-1945-1 (1), AUS-1945-2 (1), AUS-1953 (1),
AUS-1961 (1), AUS-1964 (1), AUS-1970 (1), AUS-1983 (1), AUS-1986 (1), AUS-1997
(1), AUS-2000 (1), AUS-2007 (1), AUS-2008 (1), AZE-1991 (1), AZE-1992-1 (1),
AZE-1992-2 (1), AZE-1992-3 (1), AZE-1992-4 (1), AZE-1993 (1), AZE-2003 (1),
BAH-1971 (1), BAH-1999 (1), BAR-1966 (1), BAR-1976 (1), BAR-1985 (1), BAR-1986
(1), BAR-1987 (1), BAR-1994 (1), BAR-2008 (1), BAR-2010 (1), BAV-1864 (1),
BEL-1871 (1), BEL-1878 (1), BEL-1884-1 (1), BEL-1884-2 (1), BEL-1894 (1),
BEL-1896 (1), BEL-1899-1 (1), BEL-1899-2 (1), BEL-1907 (1), BEL-1908 (1),
BEL-1911 (1), BEL-1914-1 (1), BEL-1914-2 (1), BEL-1917 (1), BEL-1918-1 (1),
BEL-1918-2 (1), BEL-1920 (1), BEL-1921 (1), BEL-1925-1 (1), BEL-1925-2 (1),
BEL-1926 (1), BEL-1931 (1), BEL-1932 (1), BEL-1934 (1), BEL-1935 (1), BEL-1937
(1), BEL-1938 (1), BEL-1939 (1), BEL-1944 (1), BEL-1945 (1), BEL-1946-1 (1),
BEL-1946-2 (1), BEL-1946-3 (1), BEL-1947 (1), BEL-1949 (1), BEL-1950-1 (1),
BEL-1950-2 (1), BEL-1952 (1), BEL-1954 (1), BEL-1958 (1), BEL-1961 (1),
BEL-1965 (1), BEL-1966 (1), BEL-1968 (1), BEL-1973 (1), BEL-1974 (1), BEL-1978
(1), BEL-1979 (1), BEL-1981-1 (1), BEL-1981-2 (1), BEL-1992 (1), BEL-1999 (1),
BEL-2008-1 (1), BEL-2008-2 (1), BEL-2009 (1), BEL-2011 (1), BEL-2014 (1),
BEN-1960 (1), BEN-1963 (1), BEN-1964 (1), BEN-1965-1 (1), BEN-1965-2 (1),
BEN-1967 (1), BEN-1968 (1), BEN-1969-1 (1), BEN-1969-2 (1), BEN-1970 (1),
BEN-1972-1 (1), BEN-1972-2 (1), BEN-1991 (1), BEN-1996 (1), BEN-2006 (1),
BFO-1960 (1), BFO-1966 (1), BFO-1971 (1), BFO-1974 (1), BFO-1980 (1), BFO-1982
(1), BFO-1983 (1), BFO-1987 (1), BFO-2014-1 (1), BFO-2014-2 (1), BFO-2014-3
(1), BHM-1967 (1), BHM-1992 (1), BHM-2002 (1), BHM-2005-1 (1), BHM-2005-2 (1),
BHM-2007 (1), BHM-2012 (1), BHU-1926 (1), BHU-1952 (1), BHU-1972 (1), BHU-1998
(1), BHU-1999 (1), BHU-2000 (1), BHU-2001 (1), BHU-2002 (1), BHU-2003 (1),
BHU-2004 (1), BHU-2005 (1), BHU-2006 (1), BHU-2007 (1), BHU-2008 (1), BHU-2013
(1), BLR-1991 (1), BLR-1994-1 (1), BLR-1994-2 (1), BLR-1994-3 (1), BLZ-1961
(1), BLZ-1984 (1), BLZ-1989 (1), BLZ-1993 (1), BLZ-1998 (1), BLZ-2008 (1),
BNG-1971 (1), BNG-1972 (1), BNG-1975-1 (1), BNG-1975-2 (1), BNG-1981 (1),
BNG-1982 (1), BNG-1990 (1), BNG-1991 (1), BNG-1996-1 (1), BNG-1996-2 (1),
BNG-2001-1 (1), BNG-2001-2 (1), BNG-2006 (1), BNG-2007 (1), BNG-2009 (1),
BOL-1874 (1), BOL-1876 (1), BOL-1880 (1), BOL-1884 (1), BOL-1888 (1), BOL-1892
(1), BOL-1896 (1), BOL-1899 (1), BOL-1904 (1), BOL-1909 (1), BOL-1913 (1),
BOL-1917 (1), BOL-1920 (1), BOL-1925 (1), BOL-1926 (1), BOL-1930 (1), BOL-1931
(1), BOL-1934 (1), BOL-1936 (1), BOL-1937 (1), BOL-1939 (1), BOL-1940 (1),
BOL-1943 (1), BOL-1946-1 (1), BOL-1946-2 (1), BOL-1947 (1), BOL-1949 (1),
BOL-1951 (1), BOL-1952-1 (1), BOL-1952-2 (1), BOL-1956 (1), BOL-1960 (1),
BOL-1964 (1), BOL-1966-1 (1), BOL-1966-2 (1), BOL-1969-1 (1), BOL-1969-2 (1),
BOL-1970 (1), BOL-1971 (1), BOL-1978-1 (1), BOL-1978-2 (1), BOL-1979-1 (1),
BOL-1979-2 (1), BOL-1979-3 (1), BOL-1980 (1), BOL-1981 (1), BOL-1982-1 (1),
BOL-1982-2 (1), BOL-1985 (1), BOL-1989 (1), BOL-1993 (1), BOL-1997 (1),
BOL-2001 (1), BOL-2002 (1), BOL-2003 (1), BOL-2005 (1), BOL-2006 (1), BOS-1990
(1), BOS-1998 (1), BOS-1999 (1), BOS-2000-1 (1), BOS-2000-2 (1), BOS-2001 (1),
BOS-2002-1 (1), BOS-2002-2 (1), BOS-2003-1 (1), BOS-2003-2 (1), BOS-2004-1 (1),
BOS-2004-2 (1), BOS-2005 (1), BOS-2006-1 (1), BOS-2006-2 (1), BOS-2007 (1),
BOS-2008-1 (1), BOS-2008-2 (1), BOS-2009 (1), BOS-2010-1 (1), BOS-2010-2 (1),
BOS-2011 (1), BOS-2014-1 (1), BOS-2014-2 (1), BOT-1966 (1), BOT-1980 (1),
BOT-1998 (1), BOT-2008 (1), BRA-1840 (1), BRA-1889 (1), BRA-1891 (1), BRA-1894
(1), BRA-1898 (1), BRA-1902 (1), BRA-1906 (1), BRA-1909 (1), BRA-1910 (1),
BRA-1914 (1), BRA-1918 (1), BRA-1919 (1), BRA-1922 (1), BRA-1926 (1), BRA-1930
(1), BRA-1945 (1), BRA-1946 (1), BRA-1951 (1), BRA-1954 (1), BRA-1955-1 (1),
BRA-1955-2 (1), BRA-1956 (1), BRA-1961-1 (1), BRA-1961-2 (1), BRA-1961-3 (1),
BRA-1964-1 (1), BRA-1964-2 (1), BRA-1967 (1), BRA-1969-1 (1), BRA-1969-2 (1),
BRA-1974 (1), BRA-1979 (1), BRA-1985 (1), BRA-1990 (1), BRA-1992 (1), BRA-1995
(1), BRA-2003 (1), BRA-2011 (1), BRU-1967 (1), BUI-1962 (1), BUI-1966-1 (1),
BUI-1966-2 (1), BUI-1976 (1), BUI-1987 (1), BUI-1993-1 (1), BUI-1993-2 (1),
BUI-1993-3 (1), BUI-1994-1 (1), BUI-1994-2 (1), BUI-1996 (1), BUI-2003 (1),
BUI-2005 (1), BUL-1887 (1), BUL-1918 (1), BUL-1943 (1), BUL-1944 (1), BUL-1946
(1), BUL-1949 (1), BUL-1950 (1), BUL-1956 (1), BUL-1989 (1), BUL-1990-1 (1),
BUL-1990-2 (1), BUL-1991 (1), BUL-1992 (1), BUL-1994 (1), BUL-1995 (1),
BUL-1997-1 (1), BUL-1997-2 (1), BUL-2001 (1), BUL-2005 (1), BUL-2009 (1),
BUL-2013-1 (1), BUL-2013-2 (1), BUL-2014-1 (1), BUL-2014-2 (1), CAM-1953 (1),
CAM-1970 (1), CAM-1975 (1), CAM-1979 (1), CAM-1985 (1), CAM-1993 (1), CAM-1997
(1), CAN-1873 (1), CAN-1878 (1), CAN-1891 (1), CAN-1892 (1), CAN-1894 (1),
CAN-1896-1 (1), CAN-1896-2 (1), CAN-1911 (1), CAN-1920 (1), CAN-1921 (1),
CAN-1926-1 (1), CAN-1926-2 (1), CAN-1930 (1), CAN-1935 (1), CAN-1948 (1),
CAN-1957 (1), CAN-1963 (1), CAN-1968 (1), CAN-1979 (1), CAN-1980 (1),
CAN-1984-1 (1), CAN-1984-2 (1), CAN-1993-1 (1), CAN-1993-2 (1), CAN-2003 (1),
CAN-2006 (1), CAO-1960 (1), CAO-1982 (1), CAP-1975 (1), CAP-1991 (1), CAP-2000
(1), CAP-2001 (1), CDI-1960 (1), CDI-1993 (1), CDI-1999 (1), CDI-2000 (1),
CDI-2011 (1), CEM-2014-1 (1), CEM-2014-2 (1), CEN-1960 (1), CEN-1966 (1),
CEN-1979 (1), CEN-1981 (1), CEN-1993 (1), CEN-2003 (1), CEN-2013 (1), CHA-1960
(1), CHA-1975 (1), CHA-1979 (1), CHA-1982 (1), CHA-1990 (1), CHL-1871 (1),
CHL-1876 (1), CHL-1881 (1), CHL-1886 (1), CHL-1891 (1), CHL-1896 (1),
CHL-1901-1 (1), CHL-1901-2 (1), CHL-1906 (1), CHL-1910-1 (1), CHL-1910-2 (1),
CHL-1911 (1), CHL-1915 (1), CHL-1920 (1), CHL-1924 (1), CHL-1925-1 (1),
CHL-1925-2 (1), CHL-1925-3 (1), CHL-1925-4 (1), CHL-1927 (1), CHL-1931-1 (1),
CHL-1931-2 (1), CHL-1931-3 (1), CHL-1931-4 (1), CHL-1932-1 (1), CHL-1932-2 (1),
CHL-1932-3 (1), CHL-1932-4 (1), CHL-1932-5 (1), CHL-1932-6 (1), CHL-1938 (1),
CHL-1941 (1), CHL-1942 (1), CHL-1946-1 (1), CHL-1946-2 (1), CHL-1946-3 (1),
CHL-1946-4 (1), CHL-1946-5 (1), CHL-1952 (1), CHL-1958 (1), CHL-1964 (1),
CHL-1970 (1), CHL-1973 (1), CHL-1990 (1), CHL-1994 (1), CHL-2000 (1), CHL-2006
(1), CHL-2010 (1), CHL-2014 (1), CHN-1861 (1), CHN-1908 (1), CHN-1911 (1),
CHN-1912-1 (1), CHN-1912-2 (1), CHN-1916 (1), CHN-1917 (1), CHN-1918 (1),
CHN-1922 (1), CHN-1923 (1), CHN-1924 (1), CHN-1926 (1), CHN-1928 (1), CHN-1937
(1), CHN-1940 (1), CHN-1944 (1), CHN-1946 (1), CHN-1949 (1), CHN-1976 (1),
CHN-1980 (1), CHN-1997 (1), CHN-2003 (1), CHN-2012 (1), COL-1874 (1), COL-1876
(1), COL-1877-1 (1), COL-1877-2 (1), COL-1878 (1), COL-1880 (1), COL-1882-1
(1), COL-1882-2 (1), COL-1884 (1), COL-1894 (1), COL-1898 (1), COL-1900 (1),
COL-1904 (1), COL-1909-1 (1), COL-1909-2 (1), COL-1910 (1), COL-1914 (1),
COL-1918 (1), COL-1921 (1), COL-1922 (1), COL-1926 (1), COL-1930 (1), COL-1934
(1), COL-1938 (1), COL-1942 (1), COL-1945 (1), COL-1946 (1), COL-1950 (1),
COL-1951 (1), COL-1953-1 (1), COL-1953-2 (1), COL-1957 (1), COL-1958 (1),
COL-1962 (1), COL-1966 (1), COL-1970 (1), COL-1974 (1), COL-1978 (1), COL-1982
(1), COL-1986 (1), COL-1990 (1), COL-1994 (1), COL-1998 (1), COL-2002 (1),
COL-2010 (1), COM-1975-1 (1), COM-1975-2 (1), COM-1978 (1), COM-1989 (1),
COM-1995-1 (1), COM-1995-2 (1), COM-1996-1 (1), COM-1996-2 (1), COM-1998 (1),
COM-1999 (1), COM-2002-1 (1), COM-2002-2 (1), COM-2006 (1), COM-2011 (1),
CON-1960 (1), CON-1963 (1), CON-1968 (1), CON-1969 (1), CON-1977 (1), CON-1979
(1), CON-1992 (1), CON-1997 (1), COS-1870 (1), COS-1882-1 (1), COS-1882-2 (1),
COS-1885 (1), COS-1889 (1), COS-1890 (1), COS-1894 (1), COS-1902 (1), COS-1906
(1), COS-1910 (1), COS-1914 (1), COS-1917 (1), COS-1919-1 (1), COS-1919-2 (1),
COS-1920 (1), COS-1924 (1), COS-1928 (1), COS-1932 (1), COS-1936 (1), COS-1940
(1), COS-1944 (1), COS-1948-1 (1), COS-1948-2 (1), COS-1949 (1), COS-1953 (1),
COS-1958 (1), COS-1962 (1), COS-1966 (1), COS-1970 (1), COS-1974 (1), COS-1978
(1), COS-1982 (1), COS-1986 (1), COS-1990 (1), COS-1994 (1), COS-1998 (1),
COS-2002 (1), COS-2006 (1), COS-2010 (1), COS-2014 (1), CRO-1990 (1), CRO-1999
(1), CRO-2000-1 (1), CRO-2000-2 (1), CRO-2010 (1), CUB-1902 (1), CUB-1909 (1),
CUB-1913 (1), CUB-1921 (1), CUB-1925 (1), CUB-1933-1 (1), CUB-1933-2 (1),
CUB-1934-1 (1), CUB-1934-2 (1), CUB-1935 (1), CUB-1936-1 (1), CUB-1936-2 (1),
CUB-1940 (1), CUB-1944 (1), CUB-1948 (1), CUB-1952 (1), CUB-1959 (1), CUB-2008
(1), CYP-1960 (1), CYP-1974-1 (1), CYP-1974-2 (1), CYP-1974-3 (1), CYP-1977
(1), CYP-1988 (1), CYP-1993 (1), CYP-2003 (1), CYP-2008 (1), CYP-2013 (1),
CZE-1918 (1), CZE-1935 (1), CZE-1938-1 (1), CZE-1938-2 (1), CZE-1945 (1),
CZE-1948 (1), CZE-1953 (1), CZE-1957 (1), CZE-1968-1 (1), CZE-1968-2 (1),
CZE-1968-3 (1), CZE-1989 (1), CZE-1992 (1), CZR-1993 (1), CZR-1997 (1),
CZR-1998 (1), CZR-2002 (1), CZR-2004 (1), CZR-2005 (1), CZR-2006 (1), CZR-2009
(1), CZR-2010 (1), CZR-2013 (1), CZR-2014 (1), DEN-1874 (1), DEN-1875 (1),
DEN-1894 (1), DEN-1897 (1), DEN-1900 (1), DEN-1901 (1), DEN-1905 (1), DEN-1908
(1), DEN-1909-1 (1), DEN-1909-2 (1), DEN-1910 (1), DEN-1913 (1), DEN-1920-1
(1), DEN-1920-2 (1), DEN-1920-3 (1), DEN-1924 (1), DEN-1926 (1), DEN-1929 (1),
DEN-1942-1 (1), DEN-1942-2 (1), DEN-1945-1 (1), DEN-1945-2 (1), DEN-1947 (1),
DEN-1950 (1), DEN-1953 (1), DEN-1955 (1), DEN-1960 (1), DEN-1962 (1), DEN-1968
(1), DEN-1971 (1), DEN-1972 (1), DEN-1973 (1), DEN-1975 (1), DEN-1982 (1),
DEN-1993 (1), DEN-2001 (1), DEN-2009 (1), DEN-2011 (1), DJI-1977 (1), DJI-1999
(1), DOM-1874 (1), DOM-1876-1 (1), DOM-1876-2 (1), DOM-1876-3 (1), DOM-1876-4
(1), DOM-1878-1 (1), DOM-1878-2 (1), DOM-1878-3 (1), DOM-1878-4 (1), DOM-1879
(1), DOM-1880 (1), DOM-1882 (1), DOM-1884 (1), DOM-1885 (1), DOM-1887 (1),
DOM-1899-1 (1), DOM-1899-2 (1), DOM-1899-3 (1), DOM-1902 (1), DOM-1903-1 (1),
DOM-1903-2 (1), DOM-1905 (1), DOM-1911 (1), DOM-1912 (1), DOM-1913 (1),
DOM-1914-1 (1), DOM-1914-2 (1), DOM-1924 (1), DOM-1930-1 (1), DOM-1930-2 (1),
DOM-1961 (1), DOM-1962-1 (1), DOM-1962-2 (1), DOM-1963-1 (1), DOM-1963-2 (1),
DOM-1963-3 (1), DOM-1965-1 (1), DOM-1965-2 (1), DOM-1965-3 (1), DOM-1965-4 (1),
DOM-1966 (1), DOM-1978 (1), DOM-1982-1 (1), DOM-1982-2 (1), DOM-1986 (1),
DOM-1996 (1), DOM-2000 (1), DOM-2004 (1), DOM-2012 (1), DRC-1960-1 (1),
DRC-1960-2 (1), DRC-1960-3 (1), DRC-1965 (1), DRC-1997 (1), DRC-2001 (1),
DRV-1945 (1), DRV-1969 (1), DRV-1986-1 (1), DRV-1986-2 (1), DRV-1991 (1),
DRV-1997 (1), DRV-2001 (1), DRV-2011 (1), ECU-1869 (1), ECU-1875 (1), ECU-1876
(1), ECU-1883-1 (1), ECU-1883-2 (1), ECU-1883-3 (1), ECU-1883-4 (1), ECU-1888
(1), ECU-1892 (1), ECU-1895 (1), ECU-1901 (1), ECU-1905 (1), ECU-1906 (1),
ECU-1911-1 (1), ECU-1911-2 (1), ECU-1911-3 (1), ECU-1912-1 (1), ECU-1912-2 (1),
ECU-1912-3 (1), ECU-1916 (1), ECU-1920 (1), ECU-1924 (1), ECU-1925 (1),
ECU-1931-1 (1), ECU-1931-2 (1), ECU-1932-1 (1), ECU-1932-2 (1), ECU-1932-3 (1),
ECU-1933 (1), ECU-1934 (1), ECU-1935-1 (1), ECU-1935-2 (1), ECU-1937 (1),
ECU-1938-1 (1), ECU-1938-2 (1), ECU-1939-1 (1), ECU-1939-2 (1), ECU-1940-1 (1),
ECU-1940-2 (1), ECU-1944-1 (1), ECU-1944-2 (1), ECU-1944-3 (1), ECU-1947-1 (1),
ECU-1947-2 (1), ECU-1947-3 (1), ECU-1948 (1), ECU-1952 (1), ECU-1956 (1),
ECU-1960 (1), ECU-1961 (1), ECU-1963 (1), ECU-1966-1 (1), ECU-1966-2 (1),
ECU-1968 (1), ECU-1972 (1), ECU-1976 (1), ECU-1979 (1), ECU-1981 (1), ECU-1984
(1), ECU-1988 (1), ECU-1992 (1), ECU-1996 (1), ECU-1997-1 (1), ECU-1997-2 (1),
ECU-1997-3 (1), ECU-1998 (1), ECU-2000 (1), ECU-2003 (1), ECU-2005 (1),
ECU-2007 (1), EGY-1922 (1), EGY-1936 (1), EGY-1952 (1), EGY-1954 (1), EGY-1970
(1), EGY-1981 (1), EGY-2011 (1), EGY-2012 (1), EGY-2013 (1), EGY-2014 (1),
EQG-1963 (1), EQG-1968 (1), EQG-1979 (1), ERI-1993 (1), EST-1918 (1),
EST-1919-1 (1), EST-1919-2 (1), EST-1920 (1), EST-1921 (1), EST-1922 (1),
EST-1923 (1), EST-1924-1 (1), EST-1924-2 (1), EST-1925 (1), EST-1927 (1),
EST-1928 (1), EST-1929 (1), EST-1931 (1), EST-1932-1 (1), EST-1932-2 (1),
EST-1932-3 (1), EST-1933-1 (1), EST-1933-2 (1), EST-1940 (1), EST-1991 (1),
EST-1992-1 (1), EST-1992-2 (1), EST-1994 (1), EST-1995 (1), EST-1997 (1),
EST-1999 (1), EST-2002 (1), EST-2003 (1), EST-2005 (1), EST-2014 (1), ETH-1871
(1), ETH-1889 (1), ETH-1911 (1), ETH-1916 (1), ETH-1930 (1), ETH-1936 (1),
ETH-1941 (1), ETH-1974-1 (1), ETH-1974-2 (1), ETH-1977 (1), ETH-1991-1 (1),
ETH-1991-2 (1), ETH-2012 (1), ETM-2002 (1), ETM-2007 (1), ETM-2012 (1),
FIN-1917 (1), FIN-1918 (1), FIN-1919 (1), FIN-1925 (1), FIN-1931 (1), FIN-1937
(1), FIN-1940 (1), FIN-1944 (1), FIN-1946 (1), FIN-1956 (1), FIN-1981 (1),
FIN-1994 (1), FIN-2000 (1), FIN-2012 (1), FJI-1970 (1), FJI-1987-1 (1),
FJI-1987-2 (1), FJI-1987-3 (1), FJI-1992 (1), FJI-1999 (1), FJI-2000-1 (1),
FJI-2000-2 (1), FJI-2000-3 (1), FJI-2006 (1), FJI-2007 (1), FRN-1874 (1),
FRN-1875 (1), FRN-1876-1 (1), FRN-1876-2 (1), FRN-1877-1 (1), FRN-1877-2 (1),
FRN-1877-3 (1), FRN-1879-1 (1), FRN-1879-2 (1), FRN-1880 (1), FRN-1881 (1),
FRN-1882-1 (1), FRN-1882-2 (1), FRN-1883-1 (1), FRN-1883-2 (1), FRN-1885 (1),
FRN-1886-1 (1), FRN-1886-2 (1), FRN-1887-1 (1), FRN-1887-2 (1), FRN-1888 (1),
FRN-1889 (1), FRN-1890 (1), FRN-1892-1 (1), FRN-1892-2 (1), FRN-1893-1 (1),
FRN-1893-2 (1), FRN-1894 (1), FRN-1895-1 (1), FRN-1895-2 (1), FRN-1896 (1),
FRN-1898-1 (1), FRN-1898-2 (1), FRN-1899 (1), FRN-1902 (1), FRN-1905 (1),
FRN-1906-1 (1), FRN-1906-2 (1), FRN-1909 (1), FRN-1911-1 (1), FRN-1911-2 (1),
FRN-1912 (1), FRN-1913-1 (1), FRN-1913-2 (1), FRN-1913-3 (1), FRN-1914 (1),
FRN-1915 (1), FRN-1917-1 (1), FRN-1917-2 (1), FRN-1917-3 (1), FRN-1920-1 (1),
FRN-1920-2 (1), FRN-1921 (1), FRN-1922 (1), FRN-1924-1 (1), FRN-1924-2 (1),
FRN-1925-1 (1), FRN-1925-2 (1), FRN-1926-1 (1), FRN-1926-2 (1), FRN-1929-1 (1),
FRN-1929-2 (1), FRN-1930-1 (1), FRN-1930-2 (1), FRN-1930-3 (1), FRN-1931 (1),
FRN-1932-1 (1), FRN-1932-2 (1), FRN-1932-3 (1), FRN-1933-1 (1), FRN-1933-2 (1),
FRN-1933-3 (1), FRN-1934-1 (1), FRN-1934-2 (1), FRN-1934-3 (1), FRN-1935-1 (1),
FRN-1935-2 (1), FRN-1936-1 (1), FRN-1936-2 (1), FRN-1937 (1), FRN-1938-1 (1),
FRN-1938-2 (1), FRN-1940-1 (1), FRN-1940-2 (1), FRN-1942 (1), FRN-1944 (1),
FRN-1946-1 (1), FRN-1946-2 (1), FRN-1946-3 (1), FRN-1947-1 (1), FRN-1947-2 (1),
FRN-1948-1 (1), FRN-1948-2 (1), FRN-1948-3 (1), FRN-1949 (1), FRN-1950-1 (1),
FRN-1950-2 (1), FRN-1951-1 (1), FRN-1951-2 (1), FRN-1952-1 (1), FRN-1952-2 (1),
FRN-1953-1 (1), FRN-1953-2 (1), FRN-1954 (1), FRN-1955 (1), FRN-1956 (1),
FRN-1957-1 (1), FRN-1957-2 (1), FRN-1958-1 (1), FRN-1958-2 (1), FRN-1969-1 (1),
FRN-1969-2 (1), FRN-1974-1 (1), FRN-1974-2 (1), FRN-1981 (1), FRN-1995 (1),
FRN-2007 (1), FRN-2012 (1), GAB-1960 (1), GAB-1964-1 (1), GAB-1964-2 (1),
GAB-1967 (1), GAB-2009-1 (1), GAB-2009-2 (1), GAM-1965 (1), GAM-1994 (1),
GDR-1946 (1), GDR-1950 (1), GDR-1971 (1), GDR-1989-1 (1), GDR-1989-2 (1),
GFR-1949 (1), GFR-1963 (1), GFR-1966 (1), GFR-1969 (1), GFR-1974 (1), GFR-1982
(1), GFR-1998 (1), GFR-2005 (1), GHA-1952 (1), GHA-1966 (1), GHA-1969-1 (1),
GHA-1969-2 (1), GHA-1972 (1), GHA-1978 (1), GHA-1979-1 (1), GHA-1979-2 (1),
GHA-1981 (1), GHA-2001 (1), GHA-2009 (1), GHA-2012 (1), GMY-1858 (1),
GMY-1888-1 (1), GMY-1888-2 (1), GMY-1918 (1), GMY-1919-1 (1), GMY-1919-2 (1),
GMY-1920-1 (1), GMY-1920-2 (1), GMY-1921 (1), GMY-1922 (1), GMY-1923-1 (1),
GMY-1923-2 (1), GMY-1925 (1), GMY-1926 (1), GMY-1928 (1), GMY-1930 (1),
GMY-1932-1 (1), GMY-1932-2 (1), GMY-1933 (1), GMY-1945 (1), GNB-1974 (1),
GNB-1980 (1), GNB-1999-1 (1), GNB-1999-2 (1), GNB-2000 (1), GNB-2003-1 (1),
GNB-2003-2 (1), GNB-2005 (1), GNB-2009-1 (1), GNB-2009-2 (1), GNB-2012-1 (1),
GNB-2012-2 (1), GNB-2012-3 (1), GNB-2014 (1), GRC-1863 (1), GRC-1913 (1),
GRC-1917 (1), GRC-1920-1 (1), GRC-1920-2 (1), GRC-1922-1 (1), GRC-1922-2 (1),
GRC-1924-1 (1), GRC-1924-2 (1), GRC-1924-3 (1), GRC-1924-4 (1), GRC-1924-5 (1),
GRC-1925 (1), GRC-1926-1 (1), GRC-1926-2 (1), GRC-1928 (1), GRC-1932-1 (1),
GRC-1932-2 (1), GRC-1932-3 (1), GRC-1933-1 (1), GRC-1933-2 (1), GRC-1933-3 (1),
GRC-1935-1 (1), GRC-1935-2 (1), GRC-1936 (1), GRC-1941-1 (1), GRC-1941-2 (1),
GRC-1944 (1), GRC-1945-1 (1), GRC-1945-2 (1), GRC-1945-3 (1), GRC-1945-4 (1),
GRC-1945-5 (1), GRC-1946-1 (1), GRC-1946-2 (1), GRC-1947-1 (1), GRC-1947-2 (1),
GRC-1947-3 (1), GRC-1949 (1), GRC-1950-1 (1), GRC-1950-2 (1), GRC-1950-3 (1),
GRC-1950-4 (1), GRC-1951 (1), GRC-1952-1 (1), GRC-1952-2 (1), GRC-1955-1 (1),
GRC-1955-2 (1), GRC-1958-1 (1), GRC-1958-2 (1), GRC-1961-1 (1), GRC-1961-2 (1),
GRC-1963-1 (1), GRC-1963-2 (1), GRC-1963-3 (1), GRC-1963-4 (1), GRC-1964 (1),
GRC-1965-1 (1), GRC-1965-2 (1), GRC-1965-3 (1), GRC-1966 (1), GRC-1967-1 (1),
GRC-1967-2 (1), GRC-1967-3 (1), GRC-1973 (1), GRC-1974 (1), GRC-1980 (1),
GRC-1981 (1), GRC-1989-1 (1), GRC-1989-2 (1), GRC-1989-3 (1), GRC-1990 (1),
GRC-1993 (1), GRC-1995 (1), GRC-1996 (1), GRC-2004 (1), GRC-2009 (1), GRC-2011
(1), GRC-2012-1 (1), GRC-2012-2 (1), GRG-1991 (1), GRG-1992-1 (1), GRG-1992-2
(1), GRG-2003 (1), GRG-2004 (1), GRG-2007 (1), GRG-2008 (1), GRG-2013 (1),
GUA-1873 (1), GUA-1885 (1), GUA-1892 (1), GUA-1898 (1), GUA-1920 (1), GUA-1921
(1), GUA-1926 (1), GUA-1930-1 (1), GUA-1930-2 (1), GUA-1931-1 (1), GUA-1931-2
(1), GUA-1944-1 (1), GUA-1944-2 (1), GUA-1945 (1), GUA-1951 (1), GUA-1954-1
(1), GUA-1954-2 (1), GUA-1954-3 (1), GUA-1957-1 (1), GUA-1957-2 (1), GUA-1957-3
(1), GUA-1958 (1), GUA-1963 (1), GUA-1966 (1), GUA-1970 (1), GUA-1974 (1),
GUA-1978 (1), GUA-1982 (1), GUA-1983 (1), GUA-1986 (1), GUA-1991 (1),
GUA-1993-1 (1), GUA-1993-2 (1), GUA-1996 (1), GUA-2000 (1), GUA-2004 (1),
GUA-2008 (1), GUA-2012 (1), GUI-1958 (1), GUI-1984-1 (1), GUI-1984-2 (1),
GUI-2008-1 (1), GUI-2008-2 (1), GUI-2009 (1), GUI-2010 (1), GUY-1966 (1),
GUY-1985 (1), GUY-1992 (1), GUY-1997-1 (1), GUY-1997-2 (1), GUY-1999 (1),
GUY-2011 (1), HAI-1874 (1), HAI-1876 (1), HAI-1879-1 (1), HAI-1879-2 (1),
HAI-1879-3 (1), HAI-1888-1 (1), HAI-1888-2 (1), HAI-1889 (1), HAI-1896 (1),
HAI-1902-1 (1), HAI-1902-2 (1), HAI-1908 (1), HAI-1911 (1), HAI-1912 (1),
HAI-1913 (1), HAI-1914-1 (1), HAI-1914-2 (1), HAI-1915 (1), HAI-1934 (1),
HAI-1941 (1), HAI-1946-1 (1), HAI-1946-2 (1), HAI-1950 (1), HAI-1956 (1),
HAI-1957-1 (1), HAI-1957-2 (1), HAI-1957-3 (1), HAI-1957-4 (1), HAI-1957-5 (1),
HAI-1957-6 (1), HAI-1957-7 (1), HAI-1971 (1), HAI-1986 (1), HAI-1988-1 (1),
HAI-1988-2 (1), HAI-1988-3 (1), HAI-1990-1 (1), HAI-1990-2 (1), HAI-1991-1 (1),
HAI-1991-2 (1), HAI-1994 (1), HAI-1996 (1), HAI-2001 (1), HAI-2004 (1),
HAI-2006 (1), HAI-2011 (1), HON-1874 (1), HON-1876 (1), HON-1883-1 (1),
HON-1883-2 (1), HON-1891 (1), HON-1893 (1), HON-1894 (1), HON-1899 (1),
HON-1903 (1), HON-1907 (1), HON-1911 (1), HON-1912 (1), HON-1913 (1), HON-1915
(1), HON-1916 (1), HON-1919 (1), HON-1920 (1), HON-1924 (1), HON-1925 (1),
HON-1929 (1), HON-1933 (1), HON-1949 (1), HON-1954 (1), HON-1956 (1), HON-1957
(1), HON-1963 (1), HON-1971 (1), HON-1972 (1), HON-1975 (1), HON-1978 (1),
HON-1982 (1), HON-1986 (1), HON-1990 (1), HON-1994 (1), HON-1998 (1), HON-2002
(1), HON-2006 (1), HON-2009 (1), HON-2010 (1), HON-2014 (1), HUN-1918 (1),
HUN-1919-1 (1), HUN-1919-2 (1), HUN-1919-3 (1), HUN-1919-4 (1), HUN-1920 (1),
HUN-1945 (1), HUN-1956-1 (1), HUN-1956-2 (1), HUN-1956-3 (1), HUN-1988 (1),
HUN-1989 (1), HUN-1990 (1), HUN-1993 (1), HUN-1994 (1), HUN-1998 (1), HUN-2002
(1), HUN-2004 (1), HUN-2009 (1), HUN-2010 (1), ICE-1944-1 (1), ICE-1944-2 (1),
ICE-1947 (1), ICE-1949 (1), ICE-1950 (1), ICE-1953 (1), ICE-1956 (1), ICE-1958
(1), ICE-1959 (1), ICE-1963 (1), ICE-1970 (1), ICE-1971 (1), ICE-1974 (1),
ICE-1978 (1), ICE-1979 (1), ICE-1980 (1), ICE-1983 (1), ICE-1987 (1), ICE-1988
(1), ICE-1991 (1), ICE-2004 (1), ICE-2006 (1), ICE-2009 (1), ICE-2013 (1),
IND-1947 (1), IND-1964-1 (1), IND-1964-2 (1), IND-1966-1 (1), IND-1966-2 (1),
IND-1977 (1), IND-1979 (1), IND-1980 (1), IND-1984 (1), IND-1989 (1), IND-1990
(1), IND-1991 (1), IND-1996-1 (1), IND-1996-2 (1), IND-1997 (1), IND-1998 (1),
IND-2004 (1), IND-2014 (1), INS-1945 (1), INS-1948 (1), INS-1949-1 (1),
INS-1949-2 (1), INS-1966 (1), INS-1998 (1), INS-1999 (1), INS-2001 (1),
INS-2004 (1), INS-2014 (1), IRE-1919 (1), IRE-1922-1 (1), IRE-1922-2 (1),
IRE-1922-3 (1), IRE-1932 (1), IRE-1948 (1), IRE-1951 (1), IRE-1954 (1),
IRE-1957 (1), IRE-1959 (1), IRE-1966 (1), IRE-1973 (1), IRE-1977 (1), IRE-1979
(1), IRE-1981 (1), IRE-1982-1 (1), IRE-1982-2 (1), IRE-1987 (1), IRE-1992 (1),
IRE-1994 (1), IRE-1997 (1), IRE-2008 (1), IRE-2011 (1), IRN-1848 (1), IRN-1896
(1), IRN-1907 (1), IRN-1909 (1), IRN-1921 (1), IRN-1941 (1), IRN-1951 (1),
IRN-1953 (1), IRN-1979 (1), IRN-1989-1 (1), IRN-1989-2 (1), IRN-1997 (1),
IRN-2005 (1), IRN-2013 (1), IRQ-1932 (1), IRQ-1933 (1), IRQ-1939 (1), IRQ-1953
(1), IRQ-1958 (1), IRQ-1963 (1), IRQ-1966 (1), IRQ-1968 (1), IRQ-1979 (1),
IRQ-2003-1 (1), IRQ-2003-2 (1), IRQ-2003-3 (1), IRQ-2004 (1), IRQ-2005 (1),
IRQ-2006 (1), IRQ-2014 (1), ISR-1948 (1), ISR-1953 (1), ISR-1955 (1), ISR-1963
(1), ISR-1969-1 (1), ISR-1969-2 (1), ISR-1974 (1), ISR-1977-1 (1), ISR-1977-2
(1), ISR-1983 (1), ISR-1984 (1), ISR-1986 (1), ISR-1992 (1), ISR-1995 (1),
ISR-1996 (1), ISR-1999 (1), ISR-2001 (1), ISR-2006 (1), ISR-2009 (1), ITA-1873
(1), ITA-1876 (1), ITA-1878-1 (1), ITA-1878-2 (1), ITA-1879 (1), ITA-1881 (1),
ITA-1887 (1), ITA-1891 (1), ITA-1892 (1), ITA-1893 (1), ITA-1896 (1), ITA-1898
(1), ITA-1900 (1), ITA-1901 (1), ITA-1903 (1), ITA-1905 (1), ITA-1906-1 (1),
ITA-1906-2 (1), ITA-1909 (1), ITA-1910 (1), ITA-1911 (1), ITA-1914 (1),
ITA-1916 (1), ITA-1917 (1), ITA-1919 (1), ITA-1920 (1), ITA-1921 (1),
ITA-1922-1 (1), ITA-1922-2 (1), ITA-1943 (1), ITA-1944 (1), ITA-1945-1 (1),
ITA-1945-2 (1), ITA-1953 (1), ITA-1954-1 (1), ITA-1954-2 (1), ITA-1955 (1),
ITA-1957 (1), ITA-1958 (1), ITA-1959 (1), ITA-1960-1 (1), ITA-1960-2 (1),
ITA-1963-1 (1), ITA-1963-2 (1), ITA-1968-1 (1), ITA-1968-2 (1), ITA-1970 (1),
ITA-1972 (1), ITA-1973 (1), ITA-1974 (1), ITA-1976 (1), ITA-1979 (1), ITA-1980
(1), ITA-1981 (1), ITA-1982 (1), ITA-1983 (1), ITA-1987-1 (1), ITA-1987-2 (1),
ITA-1988 (1), ITA-1989 (1), ITA-1992 (1), ITA-1993 (1), ITA-1994 (1), ITA-1995
(1), ITA-1996 (1), ITA-1998 (1), ITA-2000 (1), ITA-2001 (1), ITA-2006 (1),
ITA-2008 (1), ITA-2011 (1), ITA-2013 (1), ITA-2014 (1), JAM-1962 (1),
JAM-1967-1 (1), JAM-1967-2 (1), JAM-1972 (1), JAM-1980 (1), JAM-1989 (1),
JAM-1992 (1), JAM-2006 (1), JAM-2007 (1), JAM-2011 (1), JAM-2012 (1), JOR-1921
(1), JOR-1951-1 (1), JOR-1951-2 (1), JOR-1952 (1), JOR-1999 (1), JPN-1868 (1),
JPN-1885 (1), JPN-1888 (1), JPN-1889 (1), JPN-1891 (1), JPN-1892 (1), JPN-1896
(1), JPN-1897-1 (1), JPN-1897-2 (1), JPN-1898-1 (1), JPN-1898-2 (1), JPN-1898-3
(1), JPN-1900 (1), JPN-1901-1 (1), JPN-1901-2 (1), JPN-1906 (1), JPN-1908 (1),
JPN-1911 (1), JPN-1912 (1), JPN-1913 (1), JPN-1914 (1), JPN-1916 (1), JPN-1918
(1), JPN-1921-1 (1), JPN-1921-2 (1), JPN-1922 (1), JPN-1923-1 (1), JPN-1923-2
(1), JPN-1923-3 (1), JPN-1924 (1), JPN-1926 (1), JPN-1927 (1), JPN-1929 (1),
JPN-1930 (1), JPN-1931-1 (1), JPN-1931-2 (1), JPN-1932-1 (1), JPN-1932-2 (1),
JPN-1934 (1), JPN-1936 (1), JPN-1937-1 (1), JPN-1937-2 (1), JPN-1939-1 (1),
JPN-1939-2 (1), JPN-1940-1 (1), JPN-1940-2 (1), JPN-1941 (1), JPN-1944 (1),
JPN-1945-1 (1), JPN-1945-2 (1), JPN-1951 (1), JPN-1952 (1), JPN-1954 (1),
JPN-1956 (1), JPN-1957 (1), JPN-1960 (1), JPN-1964 (1), JPN-1972 (1), JPN-1974
(1), JPN-1976 (1), JPN-1978 (1), JPN-1980-1 (1), JPN-1980-2 (1), JPN-1982 (1),
JPN-1987 (1), JPN-1989-1 (1), JPN-1989-2 (1), JPN-1991 (1), JPN-1993 (1),
JPN-1994-1 (1), JPN-1994-2 (1), JPN-1996 (1), JPN-1998 (1), JPN-2000-1 (1),
JPN-2000-2 (1), JPN-2001 (1), JPN-2006 (1), JPN-2007 (1), JPN-2008 (1),
JPN-2009 (1), JPN-2010 (1), JPN-2011 (1), JPN-2012 (1), KEN-1963 (1), KEN-1978
(1), KEN-2002 (1), KEN-2013 (1), KOR-1873 (1), KOR-1907 (1), KOS-2008 (1),
KOS-2014 (1), KUW-1950 (1), KUW-1965 (1), KUW-1978 (1), KUW-1990 (1), KUW-1991
(1), KUW-2006-1 (1), KUW-2006-2 (1), KYR-1990 (1), KYR-2004 (1), KYR-2010 (1),
KYR-2011 (1), KZK-1990 (1), LAO-1951 (1), LAO-1954 (1), LAO-1956 (1), LAO-1958
(1), LAO-1959 (1), LAO-1960-1 (1), LAO-1960-2 (1), LAO-1960-3 (1), LAO-1960-4
(1), LAO-1960-5 (1), LAO-1962 (1), LAO-1964-1 (1), LAO-1964-2 (1), LAO-1975
(1), LAO-1992 (1), LAO-1998 (1), LAO-2006 (1), LAT-1918 (1), LAT-1919-1 (1),
LAT-1919-2 (1), LAT-1919-3 (1), LAT-1921 (1), LAT-1923-1 (1), LAT-1923-2 (1),
LAT-1924-1 (1), LAT-1924-2 (1), LAT-1925 (1), LAT-1926-1 (1), LAT-1926-2 (1),
LAT-1928-1 (1), LAT-1928-2 (1), LAT-1931-1 (1), LAT-1931-2 (1), LAT-1933 (1),
LAT-1934 (1), LAT-1990 (1), LAT-1993 (1), LAT-1994 (1), LAT-1995 (1), LAT-1997
(1), LAT-1998 (1), LAT-1999 (1), LAT-2000 (1), LAT-2002 (1), LAT-2004-1 (1),
LAT-2004-2 (1), LAT-2007 (1), LAT-2009 (1), LAT-2014 (1), LBR-1872 (1),
LBR-1875 (1), LBR-1876 (1), LBR-1878 (1), LBR-1883 (1), LBR-1884 (1), LBR-1892
(1), LBR-1896 (1), LBR-1900 (1), LBR-1904 (1), LBR-1912 (1), LBR-1920 (1),
LBR-1930 (1), LBR-1944 (1), LBR-1971 (1), LBR-1980 (1), LBR-1990 (1), LBR-1994
(1), LBR-1995 (1), LBR-1996 (1), LBR-1997 (1), LBR-2003-1 (1), LBR-2003-2 (1),
LBR-2006 (1), LEB-1943 (1), LEB-1952-1 (1), LEB-1952-2 (1), LEB-1958 (1),
LEB-1964 (1), LEB-1970 (1), LEB-1976 (1), LEB-1982 (1), LEB-1988 (1),
LEB-1989-1 (1), LEB-1989-2 (1), LEB-1998 (1), LEB-2007 (1), LEB-2008 (1),
LEB-2014 (1), LES-1966 (1), LES-1986 (1), LES-1991 (1), LES-1993 (1),
LES-1994-1 (1), LES-1994-2 (1), LES-1998 (1), LES-2012 (1), LIB-1951 (1),
LIB-1969 (1), LIB-2011 (1), LIB-2012-1 (1), LIB-2012-2 (1), LIB-2013-1 (1),
LIB-2013-2 (1), LIB-2014 (1), LIT-1917 (1), LIT-1918-1 (1), LIT-1918-2 (1),
LIT-1920 (1), LIT-1926-1 (1), LIT-1926-2 (1), LIT-1990 (1), LIT-1992 (1),
LIT-1998 (1), LIT-2003 (1), LIT-2004-1 (1), LIT-2004-2 (1), LIT-2009 (1),
LUX-1874 (1), LUX-1885 (1), LUX-1888 (1), LUX-1915-1 (1), LUX-1915-2 (1),
LUX-1916 (1), LUX-1917 (1), LUX-1918 (1), LUX-1925 (1), LUX-1926 (1), LUX-1937
(1), LUX-1945 (1), LUX-1953 (1), LUX-1958 (1), LUX-1959 (1), LUX-1974 (1),
LUX-1979 (1), LUX-1984 (1), LUX-1995 (1), LUX-2013 (1), MAA-1960 (1), MAA-1978
(1), MAA-1979-1 (1), MAA-1979-2 (1), MAA-1979-3 (1), MAA-1980 (1), MAA-1984
(1), MAA-2005 (1), MAA-2007 (1), MAA-2008 (1), MAA-2009-1 (1), MAA-2009-2 (1),
MAC-1991 (1), MAC-1992 (1), MAC-1998 (1), MAC-2002 (1), MAC-2004-1 (1),
MAC-2004-2 (1), MAC-2004-3 (1), MAC-2004-4 (1), MAC-2006 (1), MAD-1953 (1),
MAD-1957 (1), MAD-1978 (1), MAD-2008 (1), MAD-2012 (1), MAD-2013 (1), MAG-1864
(1), MAG-1960 (1), MAG-1972 (1), MAG-1975-1 (1), MAG-1975-2 (1), MAG-1975-3
(1), MAG-1993 (1), MAG-1996 (1), MAG-1997 (1), MAG-2002 (1), MAG-2009 (1),
MAG-2014 (1), MAL-1957 (1), MAL-1959-1 (1), MAL-1959-2 (1), MAL-1970 (1),
MAL-1976 (1), MAL-1981 (1), MAL-2003 (1), MAL-2009 (1), MAS-1968 (1), MAS-1982
(1), MAS-1995 (1), MAS-2000 (1), MAS-2003 (1), MAS-2005 (1), MAS-2014 (1),
MAW-1964 (1), MAW-1994 (1), MAW-2004 (1), MAW-2012 (1), MAW-2014 (1), MEX-1872
(1), MEX-1876 (1), MEX-1911-1 (1), MEX-1911-2 (1), MEX-1913 (1), MEX-1915 (1),
MEX-1920-1 (1), MEX-1920-2 (1), MEX-1924 (1), MEX-1928 (1), MEX-1930 (1),
MEX-1932 (1), MEX-1934 (1), MEX-1940 (1), MEX-1946 (1), MEX-1952 (1), MEX-1958
(1), MEX-1964 (1), MEX-1970 (1), MEX-1976 (1), MEX-1982 (1), MEX-1988 (1),
MEX-1994 (1), MEX-2000 (1), MEX-2006 (1), MEX-2012 (1), MLD-1990 (1), MLD-1997
(1), MLD-2001 (1), MLD-2008 (1), MLD-2009-1 (1), MLD-2009-2 (1), MLD-2013 (1),
MLI-1960 (1), MLI-1968 (1), MLI-1991 (1), MLI-1992 (1), MLI-2002 (1),
MLI-2012-1 (1), MLI-2012-2 (1), MLI-2013 (1), MLT-1962 (1), MLT-1971 (1),
MLT-1984 (1), MLT-1987 (1), MLT-1996 (1), MLT-1998 (1), MLT-2004 (1), MLT-2013
(1), MNG-1860 (1), MON-1921-1 (1), MON-1921-2 (1), MON-1922 (1), MON-1923 (1),
MON-1924 (1), MON-1925 (1), MON-1928 (1), MON-1936 (1), MON-1952 (1), MON-1984
(1), MON-1990 (1), MON-1997 (1), MON-2005 (1), MON-2009 (1), MOR-1873 (1),
MOR-1894 (1), MOR-1956 (1), MOR-1961 (1), MOR-1999 (1), MYA-1853 (1), MYA-1878
(1), MYA-1885 (1), MYA-1948 (1), MYA-1956 (1), MYA-1957 (1), MYA-1958 (1),
MYA-1960 (1), MYA-1962 (1), MYA-1988-1 (1), MYA-1988-2 (1), MYA-1988-3 (1),
MYA-1992 (1), MYA-2011 (1), MZM-1975 (1), MZM-1986 (1), MZM-2005 (1), NAM-1990
(1), NAM-2005 (1), NEP-1846 (1), NEP-1877 (1), NEP-1885 (1), NEP-1901-1 (1),
NEP-1901-2 (1), NEP-1929 (1), NEP-1932 (1), NEP-1945 (1), NEP-1948 (1),
NEP-1951 (1), NEP-1955 (1), NEP-1972 (1), NEP-1990 (1), NEP-1991 (1), NEP-1994
(1), NEP-1995 (1), NEP-1997-1 (1), NEP-1997-2 (1), NEP-1998 (1), NEP-1999 (1),
NEP-2000 (1), NEP-2001 (1), NEP-2002 (1), NEP-2003 (1), NEP-2004 (1), NEP-2005
(1), NEP-2006 (1), NEP-2008 (1), NEP-2009 (1), NEP-2011-1 (1), NEP-2011-2 (1),
NEP-2013 (1), NEP-2014 (1), NEW-1906 (1), NEW-1912-1 (1), NEW-1912-2 (1),
NEW-1925-1 (1), NEW-1925-2 (1), NEW-1928 (1), NEW-1930 (1), NEW-1935 (1),
NEW-1940 (1), NEW-1949 (1), NEW-1957-1 (1), NEW-1957-2 (1), NEW-1960 (1),
NEW-1972-1 (1), NEW-1972-2 (1), NEW-1974 (1), NEW-1975 (1), NEW-1984 (1),
NEW-1989 (1), NEW-1990-1 (1), NEW-1990-2 (1), NEW-1997 (1), NEW-1999 (1),
NEW-2008 (1), NIC-1871 (1), NIC-1875 (1), NIC-1879 (1), NIC-1883 (1), NIC-1887
(1), NIC-1889 (1), NIC-1893 (1), NIC-1909 (1), NIC-1910-1 (1), NIC-1910-2 (1),
NIC-1911 (1), NIC-1917 (1), NIC-1921 (1), NIC-1923 (1), NIC-1925 (1),
NIC-1926-1 (1), NIC-1926-2 (1), NIC-1929 (1), NIC-1933 (1), NIC-1936 (1),
NIC-1937 (1), NIC-1947-1 (1), NIC-1947-2 (1), NIC-1956 (1), NIC-1963 (1),
NIC-1966 (1), NIC-1967 (1), NIC-1979 (1), NIC-1990 (1), NIC-1997 (1), NIC-2002
(1), NIC-2007 (1), NIG-1960 (1), NIG-1966-1 (1), NIG-1966-2 (1), NIG-1975 (1),
NIG-1976 (1), NIG-1979 (1), NIG-1983 (1), NIG-1985 (1), NIG-1993-1 (1),
NIG-1993-2 (1), NIG-1998 (1), NIG-1999 (1), NIG-2007 (1), NIG-2010 (1),
NIR-1960 (1), NIR-1974 (1), NIR-1987 (1), NIR-1993 (1), NIR-1996 (1),
NIR-1999-1 (1), NIR-1999-2 (1), NIR-2010 (1), NIR-2011 (1), NOR-1905 (1),
NOR-1907 (1), NOR-1908 (1), NOR-1910 (1), NOR-1912 (1), NOR-1913 (1), NOR-1920
(1), NOR-1921 (1), NOR-1923-1 (1), NOR-1923-2 (1), NOR-1924 (1), NOR-1926 (1),
NOR-1928-1 (1), NOR-1928-2 (1), NOR-1931 (1), NOR-1932-1 (1), NOR-1932-2 (1),
NOR-1933 (1), NOR-1935 (1), NOR-1940 (1), NOR-1945-1 (1), NOR-1945-2 (1),
NOR-1951 (1), NOR-1955 (1), NOR-1963-1 (1), NOR-1963-2 (1), NOR-1965 (1),
NOR-1971 (1), NOR-1972 (1), NOR-1973 (1), NOR-1976 (1), NOR-1981-1 (1),
NOR-1981-2 (1), NOR-1986 (1), NOR-1989 (1), NOR-1990 (1), NOR-1996 (1),
NOR-1997 (1), NOR-1998-1 (1), NOR-1998-2 (1), NOR-2000 (1), NOR-2001 (1),
NOR-2005 (1), NOR-2013 (1), NTH-1874 (1), NTH-1877 (1), NTH-1879 (1), NTH-1883
(1), NTH-1888 (1), NTH-1891 (1), NTH-1894 (1), NTH-1897 (1), NTH-1901 (1),
NTH-1905 (1), NTH-1908 (1), NTH-1913 (1), NTH-1918 (1), NTH-1925 (1), NTH-1926
(1), NTH-1929 (1), NTH-1933 (1), NTH-1939 (1), NTH-1945 (1), NTH-1946 (1),
NTH-1948 (1), NTH-1958 (1), NTH-1959 (1), NTH-1963 (1), NTH-1965 (1), NTH-1966
(1), NTH-1967 (1), NTH-1971 (1), NTH-1973 (1), NTH-1977 (1), NTH-1982 (1),
NTH-1994 (1), NTH-2002 (1), NTH-2010 (1), OFS-1873 (1), OFS-1888 (1), OFS-1889
(1), OFS-1895 (1), OFS-1896 (1), OFS-1902 (1), OFS-1905 (1), OFS-1907-1 (1),
OFS-1907-2 (1), OMA-1871 (1), OMA-1888 (1), OMA-1913 (1), OMA-1932 (1),
OMA-1970 (1), PAK-1947 (1), PAK-1948 (1), PAK-1951 (1), PAK-1953 (1), PAK-1955
(1), PAK-1956 (1), PAK-1957-1 (1), PAK-1957-2 (1), PAK-1958 (1), PAK-1969 (1),
PAK-1971 (1), PAK-1977 (1), PAK-1988-1 (1), PAK-1988-2 (1), PAK-1990-1 (1),
PAK-1990-2 (1), PAK-1993-1 (1), PAK-1993-2 (1), PAK-1993-3 (1), PAK-1993-4 (1),
PAK-1996 (1), PAK-1997 (1), PAK-1999 (1), PAK-2008-1 (1), PAK-2008-2 (1),
PAK-2013 (1), PAN-1903 (1), PAN-1904 (1), PAN-1908 (1), PAN-1910-1 (1),
PAN-1910-2 (1), PAN-1912-1 (1), PAN-1912-2 (1), PAN-1912-3 (1), PAN-1916 (1),
PAN-1918-1 (1), PAN-1918-2 (1), PAN-1920-1 (1), PAN-1920-2 (1), PAN-1924 (1),
PAN-1928-1 (1), PAN-1928-2 (1), PAN-1931-1 (1), PAN-1931-2 (1), PAN-1932 (1),
PAN-1936 (1), PAN-1939 (1), PAN-1940 (1), PAN-1941-1 (1), PAN-1941-2 (1),
PAN-1945 (1), PAN-1948 (1), PAN-1949-1 (1), PAN-1949-2 (1), PAN-1949-3 (1),
PAN-1951 (1), PAN-1952 (1), PAN-1955-1 (1), PAN-1955-2 (1), PAN-1956 (1),
PAN-1960 (1), PAN-1964 (1), PAN-1968-1 (1), PAN-1968-2 (1), PAN-1981 (1),
PAN-1982 (1), PAN-1983 (1), PAN-1990 (1), PAN-1994 (1), PAN-1999 (1), PAN-2004
(1), PAN-2009 (1), PAN-2014 (1), PAR-1874 (1), PAR-1877 (1), PAR-1878 (1),
PAR-1880 (1), PAR-1886 (1), PAR-1890 (1), PAR-1894-1 (1), PAR-1894-2 (1),
PAR-1898 (1), PAR-1902-1 (1), PAR-1902-2 (1), PAR-1904 (1), PAR-1905 (1),
PAR-1906 (1), PAR-1908 (1), PAR-1910 (1), PAR-1911-1 (1), PAR-1911-2 (1),
PAR-1912-1 (1), PAR-1912-2 (1), PAR-1912-3 (1), PAR-1916 (1), PAR-1919 (1),
PAR-1920 (1), PAR-1921-1 (1), PAR-1921-2 (1), PAR-1923 (1), PAR-1924-1 (1),
PAR-1924-2 (1), PAR-1928 (1), PAR-1931 (1), PAR-1932-1 (1), PAR-1932-2 (1),
PAR-1936 (1), PAR-1937 (1), PAR-1939 (1), PAR-1940 (1), PAR-1948-1 (1),
PAR-1948-2 (1), PAR-1949-1 (1), PAR-1949-2 (1), PAR-1949-3 (1), PAR-1954-1 (1),
PAR-1954-2 (1), PAR-1989 (1), PAR-1993 (1), PAR-1998 (1), PAR-1999 (1),
PAR-2003 (1), PAR-2008 (1), PAR-2012 (1), PAR-2013 (1), PER-1872 (1), PER-1876
(1), PER-1879-1 (1), PER-1879-2 (1), PER-1881-1 (1), PER-1881-2 (1), PER-1882
(1), PER-1885 (1), PER-1890 (1), PER-1894-1 (1), PER-1894-2 (1), PER-1895-1
(1), PER-1895-2 (1), PER-1899 (1), PER-1903 (1), PER-1904-1 (1), PER-1904-2
(1), PER-1908 (1), PER-1912 (1), PER-1914 (1), PER-1915 (1), PER-1919 (1),
PER-1930-1 (1), PER-1930-2 (1), PER-1931-1 (1), PER-1931-2 (1), PER-1931-3 (1),
PER-1931-4 (1), PER-1933 (1), PER-1939 (1), PER-1945 (1), PER-1948 (1),
PER-1950-1 (1), PER-1950-2 (1), PER-1956 (1), PER-1962 (1), PER-1963-1 (1),
PER-1963-2 (1), PER-1968 (1), PER-1975 (1), PER-1980 (1), PER-1985 (1),
PER-1990 (1), PER-2000 (1), PER-2001 (1), PER-2006 (1), PER-2011 (1), PHI-1946
(1), PHI-1948 (1), PHI-1953 (1), PHI-1957 (1), PHI-1961 (1), PHI-1965 (1),
PHI-1986 (1), PHI-1992 (1), PHI-1998 (1), PHI-2001 (1), PHI-2010 (1), PNG-1975
(1), PNG-1980 (1), PNG-1982 (1), PNG-1985 (1), PNG-1988 (1), PNG-1992 (1),
PNG-1994 (1), PNG-1997 (1), PNG-1999 (1), PNG-2002 (1), PNG-2011 (1),
POL-1918-1 (1), POL-1918-2 (1), POL-1922-1 (1), POL-1922-2 (1), POL-1926 (1),
POL-1935 (1), POL-1944 (1), POL-1956-1 (1), POL-1956-2 (1), POL-1970 (1),
POL-1980 (1), POL-1981 (1), POL-1990 (1), POL-1995 (1), POL-2005 (1),
POL-2010-1 (1), POL-2010-2 (1), POL-2010-3 (1), POL-2010-4 (1), POR-1861 (1),
POR-1889 (1), POR-1908 (1), POR-1910 (1), POR-1911 (1), POR-1915-1 (1),
POR-1915-2 (1), POR-1917 (1), POR-1918 (1), POR-1919 (1), POR-1923 (1),
POR-1925 (1), POR-1926 (1), POR-1932 (1), POR-1968 (1), POR-1974-1 (1),
POR-1974-2 (1), POR-1976 (1), POR-1986 (1), POR-1996 (1), POR-2006 (1),
PRK-1948 (1), PRK-1994 (1), PRK-2011 (1), QAT-1971 (1), QAT-1972 (1), QAT-1995
(1), QAT-2013 (1), ROK-1948 (1), ROK-1960-1 (1), ROK-1960-2 (1), ROK-1961-1
(1), ROK-1961-2 (1), ROK-1979 (1), ROK-1980-1 (1), ROK-1980-2 (1), ROK-1988
(1), ROK-1993 (1), ROK-1998 (1), ROK-2003 (1), ROK-2008 (1), ROK-2013 (1),
RUM-1866 (1), RUM-1914 (1), RUM-1927 (1), RUM-1930 (1), RUM-1940 (1), RUM-1944
(1), RUM-1947 (1), RUM-1965 (1), RUM-1989 (1), RUM-1991 (1), RUM-1992 (1),
RUM-1996 (1), RUM-1998-1 (1), RUM-1998-2 (1), RUM-1999-1 (1), RUM-1999-2 (1),
RUM-2000 (1), RUM-2004-1 (1), RUM-2004-2 (1), RUM-2007-1 (1), RUM-2007-2 (1),
RUM-2012-1 (1), RUM-2012-2 (1), RUS-1855 (1), RUS-1881 (1), RUS-1894 (1),
RUS-1917 (1), RUS-1923 (1), RUS-1953-1 (1), RUS-1953-2 (1), RUS-1964 (1),
RUS-1982 (1), RUS-1984 (1), RUS-1985 (1), RUS-1991 (1), RUS-2000 (1),
RVN-1954-1 (1), RVN-1954-2 (1), RVN-1954-3 (1), RVN-1963 (1), RVN-1964-1 (1),
RVN-1964-2 (1), RVN-1965-1 (1), RVN-1965-2 (1), RVN-1965-3 (1), RVN-1975 (1),
RWA-1961 (1), RWA-1973 (1), RWA-1994-1 (1), RWA-1994-2 (1), SAF-1910 (1),
SAF-1919 (1), SAF-1924 (1), SAF-1939 (1), SAF-1948 (1), SAF-1954 (1),
SAF-1958-1 (1), SAF-1958-2 (1), SAF-1966-1 (1), SAF-1966-2 (1), SAF-1978 (1),
SAF-1989-1 (1), SAF-1989-2 (1), SAF-1989-3 (1), SAF-1994 (1), SAF-1999 (1),
SAF-2008 (1), SAF-2009 (1), SAL-1871 (1), SAL-1876-1 (1), SAL-1876-2 (1),
SAL-1885 (1), SAL-1890 (1), SAL-1894 (1), SAL-1898 (1), SAL-1903 (1), SAL-1907
(1), SAL-1911 (1), SAL-1913 (1), SAL-1914 (1), SAL-1915 (1), SAL-1918 (1),
SAL-1919 (1), SAL-1923 (1), SAL-1927 (1), SAL-1931-1 (1), SAL-1931-2 (1),
SAL-1934 (1), SAL-1935 (1), SAL-1944-1 (1), SAL-1944-2 (1), SAL-1945 (1),
SAL-1948 (1), SAL-1949-1 (1), SAL-1949-2 (1), SAL-1950 (1), SAL-1956 (1),
SAL-1960 (1), SAL-1961 (1), SAL-1962-1 (1), SAL-1962-2 (1), SAL-1967 (1),
SAL-1972 (1), SAL-1977 (1), SAL-1979 (1), SAL-1980 (1), SAL-1982 (1), SAL-1984
(1), SAL-1989 (1), SAL-1994 (1), SAL-1999 (1), SAL-2004 (1), SAL-2009 (1),
SAL-2014 (1), SAU-1927 (1), SAU-1953 (1), SAU-1964 (1), SAU-1975 (1), SAU-1982
(1), SAU-1996 (1), SEN-1960 (1), SEN-1981 (1), SEN-2000 (1), SEN-2012 (1),
SER-1882 (1), SER-1889 (1), SER-1893 (1), SER-1903 (1), SER-1914 (1), SER-2006
(1), SER-2008 (1), SER-2012 (1), SER-2014 (1), SIE-1961 (1), SIE-1964 (1),
SIE-1967-1 (1), SIE-1967-2 (1), SIE-1968 (1), SIE-1985 (1), SIE-1992 (1),
SIE-1996-1 (1), SIE-1996-2 (1), SIE-1997 (1), SIE-1998 (1), SIE-2007 (1),
SIN-1959 (1), SIN-1990 (1), SIN-2004 (1), SLO-1993 (1), SLO-1994-1 (1),
SLO-1994-2 (1), SLO-1998 (1), SLO-2006 (1), SLO-2010 (1), SLO-2012 (1),
SLV-1990 (1), SLV-1991 (1), SLV-1992 (1), SLV-2000-1 (1), SLV-2000-2 (1),
SLV-2002 (1), SLV-2004 (1), SLV-2008 (1), SLV-2012 (1), SLV-2013 (1), SLV-2014
(1), SOL-1978 (1), SOL-1981 (1), SOL-1984 (1), SOL-1986 (1), SOL-1989 (1),
SOL-1993 (1), SOL-1994 (1), SOL-1997 (1), SOL-2000 (1), SOL-2001 (1),
SOL-2006-1 (1), SOL-2006-2 (1), SOL-2007 (1), SOL-2010 (1), SOL-2011 (1),
SOL-2014 (1), SOM-1960 (1), SOM-1967 (1), SOM-1969 (1), SOM-2012 (1), SPN-1874
(1), SPN-1875-1 (1), SPN-1875-2 (1), SPN-1879-1 (1), SPN-1879-2 (1), SPN-1881
(1), SPN-1883 (1), SPN-1884 (1), SPN-1885 (1), SPN-1890 (1), SPN-1892 (1),
SPN-1895 (1), SPN-1897-1 (1), SPN-1897-2 (1), SPN-1899 (1), SPN-1900 (1),
SPN-1901 (1), SPN-1902 (1), SPN-1903-1 (1), SPN-1903-2 (1), SPN-1904 (1),
SPN-1905-1 (1), SPN-1905-2 (1), SPN-1905-3 (1), SPN-1906-1 (1), SPN-1906-2 (1),
SPN-1906-3 (1), SPN-1907 (1), SPN-1909 (1), SPN-1910 (1), SPN-1912-1 (1),
SPN-1912-2 (1), SPN-1913 (1), SPN-1915 (1), SPN-1917-1 (1), SPN-1917-2 (1),
SPN-1917-3 (1), SPN-1918-1 (1), SPN-1918-2 (1), SPN-1918-3 (1), SPN-1919-1 (1),
SPN-1919-2 (1), SPN-1919-3 (1), SPN-1920 (1), SPN-1921-1 (1), SPN-1921-2 (1),
SPN-1921-3 (1), SPN-1922-1 (1), SPN-1922-2 (1), SPN-1923 (1), SPN-1930 (1),
SPN-1931-1 (1), SPN-1931-2 (1), SPN-1931-3 (1), SPN-1933-1 (1), SPN-1933-2 (1),
SPN-1933-3 (1), SPN-1934-1 (1), SPN-1934-2 (1), SPN-1935-1 (1), SPN-1935-2 (1),
SPN-1936-1 (1), SPN-1936-2 (1), SPN-1936-3 (1), SPN-1936-4 (1), SPN-1936-5 (1),
SPN-1937 (1), SPN-1939-1 (1), SPN-1939-2 (1), SPN-1975-1 (1), SPN-1975-2 (1),
SPN-1976 (1), SPN-1981 (1), SPN-1982 (1), SPN-1996 (1), SPN-2004 (1), SPN-2011
(1), SRI-1948 (1), SRI-1952 (1), SRI-1953 (1), SRI-1956 (1), SRI-1959 (1),
SRI-1960-1 (1), SRI-1960-2 (1), SRI-1965 (1), SRI-1970 (1), SRI-1977 (1),
SRI-1989 (1), SRI-1993 (1), SRI-1994 (1), SRI-2005 (1), SSD-2011 (1),
SUD-1956-1 (1), SUD-1956-2 (1), SUD-1958 (1), SUD-1964 (1), SUD-1965 (1),
SUD-1966 (1), SUD-1967 (1), SUD-1969 (1), SUD-1971-1 (1), SUD-1971-2 (1),
SUD-1985 (1), SUD-1986 (1), SUD-1989 (1), SUR-1975 (1), SUR-1980 (1), SUR-1988
(1), SUR-1990-1 (1), SUR-1990-2 (1), SUR-1991 (1), SUR-1996 (1), SUR-2000 (1),
SUR-2010 (1), SWA-1968 (1), SWA-1982 (1), SWA-1983 (1), SWA-1986 (1), SWD-1872
(1), SWD-1876 (1), SWD-1880 (1), SWD-1883 (1), SWD-1884 (1), SWD-1888 (1),
SWD-1889 (1), SWD-1891 (1), SWD-1900 (1), SWD-1902 (1), SWD-1905-1 (1),
SWD-1905-2 (1), SWD-1905-3 (1), SWD-1906 (1), SWD-1911 (1), SWD-1914 (1),
SWD-1917-1 (1), SWD-1917-2 (1), SWD-1920-1 (1), SWD-1920-2 (1), SWD-1921-1 (1),
SWD-1921-2 (1), SWD-1923 (1), SWD-1924 (1), SWD-1925 (1), SWD-1926 (1),
SWD-1928 (1), SWD-1930 (1), SWD-1932-1 (1), SWD-1932-2 (1), SWD-1936-1 (1),
SWD-1936-2 (1), SWD-1946 (1), SWD-1969 (1), SWD-1976 (1), SWD-1978 (1),
SWD-1979 (1), SWD-1982 (1), SWD-1986 (1), SWD-1991 (1), SWD-1994 (1), SWD-1996
(1), SWD-2006 (1), SWD-2014 (1), SWZ-1875 (1), SWZ-1876 (1), SWZ-1877 (1),
SWZ-1878 (1), SWZ-1879 (1), SWZ-1880 (1), SWZ-1881 (1), SWZ-1882 (1), SWZ-1883
(1), SWZ-1884 (1), SWZ-1885 (1), SWZ-1886 (1), SWZ-1887 (1), SWZ-1888-1 (1),
SWZ-1888-2 (1), SWZ-1890 (1), SWZ-1891 (1), SWZ-1892 (1), SWZ-1893 (1),
SWZ-1894 (1), SWZ-1895 (1), SWZ-1896 (1), SWZ-1897 (1), SWZ-1898 (1), SWZ-1899
(1), SWZ-1900 (1), SWZ-1901 (1), SWZ-1902 (1), SWZ-1903 (1), SWZ-1904 (1),
SWZ-1905 (1), SWZ-1906 (1), SWZ-1907 (1), SWZ-1908 (1), SWZ-1909 (1), SWZ-1910
(1), SWZ-1911 (1), SWZ-1912 (1), SWZ-1913 (1), SWZ-1914 (1), SWZ-1915 (1),
SWZ-1916 (1), SWZ-1917 (1), SWZ-1918 (1), SWZ-1919 (1), SWZ-1920 (1), SWZ-1921
(1), SWZ-1922 (1), SWZ-1923 (1), SWZ-1924 (1), SWZ-1925 (1), SWZ-1926 (1),
SWZ-1927 (1), SWZ-1928 (1), SWZ-1929 (1), SWZ-1930 (1), SWZ-1931 (1), SWZ-1932
(1), SWZ-1933 (1), SWZ-1934 (1), SWZ-1935 (1), SWZ-1936 (1), SWZ-1937 (1),
SWZ-1938 (1), SWZ-1939 (1), SWZ-1940 (1), SWZ-1941 (1), SWZ-1942 (1), SWZ-1943
(1), SWZ-1944 (1), SWZ-1945 (1), SWZ-1946 (1), SWZ-1947 (1), SWZ-1948 (1),
SWZ-1949 (1), SWZ-1950 (1), SWZ-1951 (1), SWZ-1952 (1), SWZ-1953 (1), SWZ-1954
(1), SWZ-1955 (1), SWZ-1956 (1), SWZ-1957 (1), SWZ-1958 (1), SWZ-1959 (1),
SWZ-1960 (1), SWZ-1961 (1), SWZ-1962 (1), SWZ-1963 (1), SWZ-1964 (1), SWZ-1965
(1), SWZ-1966 (1), SWZ-1967 (1), SWZ-1968 (1), SWZ-1969 (1), SWZ-1970 (1),
SWZ-1971 (1), SWZ-1972 (1), SWZ-1973 (1), SWZ-1974 (1), SWZ-1975 (1), SWZ-1976
(1), SWZ-1977 (1), SWZ-1978 (1), SWZ-1979 (1), SWZ-1980 (1), SWZ-1981 (1),
SWZ-1982 (1), SWZ-1983 (1), SWZ-1984 (1), SWZ-1985 (1), SWZ-1986 (1), SWZ-1987
(1), SWZ-1988 (1), SWZ-1989 (1), SWZ-1990 (1), SWZ-1991 (1), SWZ-1992 (1),
SWZ-1993 (1), SWZ-1994 (1), SWZ-1995 (1), SWZ-1996 (1), SWZ-1997 (1), SWZ-1998
(1), SWZ-1999 (1), SWZ-2000 (1), SWZ-2001 (1), SWZ-2002 (1), SWZ-2003 (1),
SWZ-2004 (1), SWZ-2005 (1), SWZ-2006 (1), SWZ-2007 (1), SWZ-2008 (1), SWZ-2009
(1), SWZ-2010 (1), SWZ-2011 (1), SWZ-2012 (1), SWZ-2013 (1), SWZ-2014 (1),
SYR-1943 (1), SYR-1949-1 (1), SYR-1949-2 (1), SYR-1949-3 (1), SYR-1954 (1),
SYR-1955 (1), SYR-1958 (1), SYR-1961-1 (1), SYR-1961-2 (1), SYR-1962-1 (1),
SYR-1962-2 (1), SYR-1963-1 (1), SYR-1963-2 (1), SYR-1966 (1), SYR-1970 (1),
SYR-1971 (1), SYR-2000 (1), TAJ-1991-1 (1), TAJ-1991-2 (1), TAJ-1992-1 (1),
TAJ-1992-2 (1), TAW-1949 (1), TAW-1950 (1), TAW-1975 (1), TAW-1978 (1),
TAW-1988 (1), TAW-2000 (1), TAW-2008 (1), TAZ-1961 (1), TAZ-1985 (1), TAZ-1995
(1), TAZ-2005 (1), TBT-1875 (1), TBT-1886 (1), TBT-1895 (1), TBT-1933 (1),
TBT-1941 (1), TBT-1950 (1), THI-1868 (1), THI-1910 (1), THI-1925 (1), THI-1932
(1), THI-1933 (1), THI-1938 (1), THI-1944 (1), THI-1946 (1), THI-1947 (1),
THI-1957-1 (1), THI-1957-2 (1), THI-1958 (1), THI-1963 (1), THI-1973 (1),
THI-1975-1 (1), THI-1975-2 (1), THI-1976-1 (1), THI-1976-2 (1), THI-1976-3 (1),
THI-1977-1 (1), THI-1977-2 (1), THI-1980 (1), THI-1988 (1), THI-1991 (1),
THI-1992-1 (1), THI-1992-2 (1), THI-1992-3 (1), THI-1995 (1), THI-1996 (1),
THI-1997 (1), THI-2001 (1), THI-2006 (1), THI-2008-1 (1), THI-2008-2 (1),
THI-2008-3 (1), THI-2008-4 (1), THI-2011 (1), THI-2014-1 (1), THI-2014-2 (1),
TKM-1990 (1), TKM-2006 (1), TOG-1960 (1), TOG-1963 (1), TOG-1967-1 (1),
TOG-1967-2 (1), TOG-2005-1 (1), TOG-2005-2 (1), TOG-2005-3 (1), TRA-1872 (1),
TRA-1877 (1), TRA-1879 (1), TRA-1881 (1), TRA-1883 (1), TRA-1900 (1), TRA-1902
(1), TRA-1905 (1), TRA-1907 (1), TRI-1962 (1), TRI-1981 (1), TRI-1986 (1),
TRI-1991 (1), TRI-1995 (1), TRI-2001 (1), TRI-2010 (1), TUN-1859 (1), TUN-1943
(1), TUN-1957 (1), TUN-1987 (1), TUN-2011-1 (1), TUN-2011-2 (1), TUN-2011-3
(1), TUR-1861 (1), TUR-1876-1 (1), TUR-1876-2 (1), TUR-1909 (1), TUR-1913 (1),
TUR-1918 (1), TUR-1922 (1), TUR-1938 (1), TUR-1950 (1), TUR-1960 (1), TUR-1961
(1), TUR-1965-1 (1), TUR-1965-2 (1), TUR-1971 (1), TUR-1972 (1), TUR-1973 (1),
TUR-1974-1 (1), TUR-1974-2 (1), TUR-1975 (1), TUR-1977-1 (1), TUR-1977-2 (1),
TUR-1978 (1), TUR-1979 (1), TUR-1980 (1), TUR-1983 (1), TUR-1989 (1),
TUR-1991-1 (1), TUR-1991-2 (1), TUR-1993-1 (1), TUR-1993-2 (1), TUR-1996-1 (1),
TUR-1996-2 (1), TUR-1997 (1), TUR-1999 (1), TUR-2002 (1), TUR-2003 (1),
TUR-2014 (1), UAE-1971 (1), UAE-2004 (1), UGA-1962 (1), UGA-1971 (1),
UGA-1979-1 (1), UGA-1979-2 (1), UGA-1980-1 (1), UGA-1980-2 (1), UGA-1985 (1),
UGA-1986 (1), UKG-1874 (1), UKG-1880 (1), UKG-1885 (1), UKG-1886-1 (1),
UKG-1886-2 (1), UKG-1892 (1), UKG-1894 (1), UKG-1895 (1), UKG-1902 (1),
UKG-1905 (1), UKG-1908 (1), UKG-1916 (1), UKG-1922 (1), UKG-1923 (1),
UKG-1924-1 (1), UKG-1924-2 (1), UKG-1929 (1), UKG-1935 (1), UKG-1937 (1),
UKG-1940 (1), UKG-1945 (1), UKG-1951 (1), UKG-1955 (1), UKG-1957 (1), UKG-1963
(1), UKG-1964 (1), UKG-1970 (1), UKG-1974 (1), UKG-1976 (1), UKG-1979 (1),
UKG-1990 (1), UKG-1997 (1), UKG-2007 (1), UKG-2010 (1), UKR-1990 (1), UKR-1994
(1), UKR-2005 (1), UKR-2010 (1), URU-1873 (1), URU-1875 (1), URU-1876 (1),
URU-1879-1 (1), URU-1879-2 (1), URU-1880 (1), URU-1882 (1), URU-1886-1 (1),
URU-1886-2 (1), URU-1886-3 (1), URU-1890 (1), URU-1894 (1), URU-1897 (1),
URU-1903 (1), URU-1907 (1), URU-1911 (1), URU-1915 (1), URU-1919 (1), URU-1923
(1), URU-1927 (1), URU-1931 (1), URU-1938 (1), URU-1943 (1), URU-1947-1 (1),
URU-1947-2 (1), URU-1951 (1), URU-1955 (1), URU-1956 (1), URU-1957 (1),
URU-1958 (1), URU-1959 (1), URU-1960 (1), URU-1961 (1), URU-1962 (1), URU-1963
(1), URU-1964 (1), URU-1965 (1), URU-1966 (1), URU-1967-1 (1), URU-1967-2 (1),
URU-1972 (1), URU-1976-1 (1), URU-1976-2 (1), URU-1981 (1), URU-1985 (1),
URU-1990 (1), URU-1995 (1), URU-2000 (1), URU-2005 (1), URU-2010 (1), USA-1869
(1), USA-1877 (1), USA-1881-1 (1), USA-1881-2 (1), USA-1885 (1), USA-1889 (1),
USA-1893 (1), USA-1897 (1), USA-1901 (1), USA-1909 (1), USA-1913 (1), USA-1921
(1), USA-1923 (1), USA-1929 (1), USA-1933 (1), USA-1945 (1), USA-1953 (1),
USA-1961 (1), USA-1963 (1), USA-1969 (1), USA-1974 (1), USA-1977 (1), USA-1981
(1), USA-1989 (1), USA-1993 (1), USA-2001 (1), USA-2009 (1), UZB-1990 (1),
VEN-1870 (1), VEN-1877-1 (1), VEN-1877-2 (1), VEN-1878 (1), VEN-1879 (1),
VEN-1884 (1), VEN-1886-1 (1), VEN-1886-2 (1), VEN-1887 (1), VEN-1888 (1),
VEN-1890 (1), VEN-1892-1 (1), VEN-1892-2 (1), VEN-1892-3 (1), VEN-1894-1 (1),
VEN-1894-2 (1), VEN-1898-1 (1), VEN-1898-2 (1), VEN-1899 (1), VEN-1908 (1),
VEN-1935 (1), VEN-1941 (1), VEN-1945 (1), VEN-1948-1 (1), VEN-1948-2 (1),
VEN-1950 (1), VEN-1958-1 (1), VEN-1958-2 (1), VEN-1959 (1), VEN-1964 (1),
VEN-1969 (1), VEN-1974 (1), VEN-1979 (1), VEN-1984 (1), VEN-1989 (1), VEN-1993
(1), VEN-1994 (1), VEN-1999 (1), VEN-2012 (1), VIE-1883 (1), VNM-1847 (1),
VNM-1883-1 (1), VNM-1883-2 (1), VNM-1885 (1), VNM-1889 (1), YAR-2012 (1),
YEM-1904 (1), YEM-1948-1 (1), YEM-1948-2 (1), YEM-1962-1 (1), YEM-1962-2 (1),
YEM-1967 (1), YEM-1974 (1), YEM-1977 (1), YEM-1978 (1), YPR-1967 (1), YPR-1969
(1), YPR-1978-1 (1), YPR-1978-2 (1), YPR-1980 (1), YPR-1986 (1), YUG-1918 (1),
YUG-1934 (1), YUG-1941 (1), YUG-1945 (1), YUG-1980-1 (1), YUG-1980-2 (1),
YUG-1981 (1), YUG-1982 (1), YUG-1983 (1), YUG-1984 (1), YUG-1985 (1), YUG-1986
(1), YUG-1987 (1), YUG-1988 (1), YUG-1989 (1), YUG-2000 (1), YUG-2003 (1),
ZAM-1964 (1), ZAM-1991 (1), ZAM-2002 (1), ZAM-2008 (1), ZAM-2011 (1), ZAM-2014
(1), ZAN-1961 (1), ZAN-1964 (1), ZIM-1965 (1), ZIM-1979 (1), ZIM-1980 (1). NAs
= 0.}


\item{leadid}{Leader ID. Factor with 2768 levels. Most common:
19b45b38-3bbd-11e5-afeb-eb6f07f9fec7 (5), 3641c628-3bbd-11e5-afeb-eb6f07f9fec7
(5), 81f7a6c8-1e42-11e4-b4cd-db5882bf8def (5),
820c40a4-1e42-11e4-b4cd-db5882bf8def (6), 820e59cc-1e42-11e4-b4cd-db5882bf8def
(5), 821072f1-1e42-11e4-b4cd-db5882bf8def (6),
8210a3c6-1e42-11e4-b4cd-db5882bf8def (5), 82113641-1e42-11e4-b4cd-db5882bf8def
(5), 821875e4-1e42-11e4-b4cd-db5882bf8def (5),
8219393a-1e42-11e4-b4cd-db5882bf8def (5), 821ca82f-1e42-11e4-b4cd-db5882bf8def
(5), 821cd90a-1e42-11e4-b4cd-db5882bf8def (5),
8220a9b1-1e42-11e4-b4cd-db5882bf8def (6), 82216d01-1e42-11e4-b4cd-db5882bf8def
(5), 823d467f-1e42-11e4-b4cd-db5882bf8def (5),
824b640c-1e42-11e4-b4cd-db5882bf8def (5). NAs = 0.}


\item{startdate}{Date at which leader entered office. Other.}


\item{enddate}{Date at which leader left office. Other.}


\item{entry}{Mode of entry. Factor with 4 levels. Most common: Foreign
Imposition (40), Irregular (571), Regular (2768), Unknown (2). NAs = 0.}


\item{exit}{Mode of exit. Factor with 8 levels. Most common: Foreign (72),
Irregular (601), Natural Death (200), Regular (2266), Retired Due to Ill Health
(66), Still in Office (170), Suicide (3), Unknown (3). NAs = 0.}


\item{exitcode}{More detailed mode of exit. Factor with 15 levels. Most common:
Regular (2537), Removed by Military, without Foreign Support (277), Removed by
Other Government Actors, without Foreign Support (42), Removed by Rebels,
without Foreign Support (41), Removed through Threat of Foreign Force (45),
Still in Office (170), Unknown (144). NAs = 0.}


\item{prevtimesinoffice}{Number of previous times in office. Numeric. Max = 6,
min = 0, distinct = 7, mean = 0.259, sd = 0.644, NAs = 0.}


\item{posttenurefate}{Fate after office. Factor with 9 levels. Most common:
Death (117), Exile (396), Imprisonment (170), Leader Still in Office (171),
Missing: Natural Death within Six Months of Losing Office (230), Missing: No
Information Found (47), OK (2224). NAs = 0.}


\item{gender}{Gender. Factor with 2 levels. Most common: F (74), M (3307). NAs
= 0.}


\item{yrborn}{Year of birth, if known. Numeric. Max = 1983, min = 1795,
distinct = 174, mean = 1900.904, sd = 40.095, NAs = 38.}


\item{yrdied}{Year of death, if known. Numeric. Max = 2013, min = -777,
distinct = 141, mean = 1160.82, sd = 1240.649, NAs = 101.}


\item{borndate}{Date of birth, if known. Other.}


\item{deathdate}{Date of death, if known. Other.}


\item{GWc}{Gleditsch-Ward alphabetic country code. See Gleditsch and Ward
(1999). Character with 188 distinct values. Most common: FRN (118), JPN (83),
SWZ (140). NAs = 0.}


\item{region}{Region. Character with 21 distinct values. Most common: South
America (512), Southern Europe (364), Western Europe (429). NAs = 0.}


\item{continent}{Continent. Character with 5 distinct values. Most common:
Americas (986), Asia (649), Europe (1237). NAs = 0.}


\item{microstate}{Indicator of whether the state is a microstate, according to
Gleditsch's list of microstates. Logical. TRUE = 0, FALSE = 3381, NAs = 0.}


\item{lon}{Longitude. Numeric. Max = 178.065, min = -115.271, distinct = 187,
mean = 3.765, sd = 64.626, NAs = 0.}


\item{lat}{Latitude. Numeric. Max = 64.963, min = -40.901, distinct = 187, mean
= 22.756, sd = 26.772, NAs = 0.}


}
}
\references{
Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate System Membership:
A Revised List of the Independent States since 1816." International
Interactions 25: 393-413. The list can be found at
\url{http://privatewww.essex.ac.uk/~ksg/statelist.html}

Goemans, Henk, Kristian Gleditsch, and Giacomo Chiozza (2009)."Introducing
Archigos: A Dataset of Political Leaders." Journal of Peace Research 46 (2):
269-83.
}
\keyword{datasets}

