% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthoritarianismBook.R
\docType{package}
\name{AuthoritarianismBook}
\alias{AuthoritarianismBook}
\alias{AuthoritarianismBook-package}
\title{AuthoritarianismBook}
\description{
A package with replication files for the book
Non-democratic Politics: Authoritarianism, Dictatorship, and Democratization
(Palgrave Macmillan, 2016), by Xavier Marquez. It is meant to make it
possible for students to easily replicate all the figures in the book and to
play with the data used to create them.
}
\details{
The code is in the form of vignettes, one for each chapter. Use
\code{browseVignettes("AuthoritarianismBook")} to view available vignettes.
}

