% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/democracy_mentions_yearly_doc.R
\docType{data}
\name{democracy_mentions_yearly}
\alias{democracy_mentions_yearly}
\title{Mentions of democracy in constitutional documents, V. 1.0}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 19392 rows and 21 columns.}
\usage{
democracy_mentions_yearly
}
\description{
A dataset of mentions of democracy in the constitutional documents collected by
the Comparative Constitutions Project. A similar variable is found in the much
larger dataset maintained by the CCP at
\url{http://comparativeconstitutionsproject.org/ccp-visualizations/}
}
\section{Variable descriptions}{


\describe{ \item{country_name}{Standardized country name. This is the same
across all datasets in this package, so you can always join them by
country_name and year. Character with 213 distinct values. Most common: Canada
(251), United Kingdom (799), United States of America (233). NAs = 0.}


\item{GWn}{Gleditsch-Ward numeric country code. See Gleditsch and Ward (1999).
Numeric. Max = 990, min = 2, distinct = 213, mean = 389.749, sd = 260.263, NAs
= 0.}


\item{year}{Year. Numeric. Max = 2013, min = 1215, distinct = 799, mean =
1925.81, sd = 98.455, NAs = 0.}


\item{constitution_name}{Name of constitution in the CCP corpus. Character with
891 distinct values. Most common: Norway's Constitution of 1814 (200), United
Kingdom's Constitution of 1215 (799), United States Of America's Constituti...
(233). NAs = 1853.}


\item{const_num}{Constitution number in the country's constitutional history.
(0 = pre-constitutional period, 1 = first constitution, etc.). Numeric. Max =
33, min = 0, distinct = 34, mean = 3.572, sd = 4.314, NAs = 0.}


\item{has_constitution}{Does the country have a constitution? (May be
suspended). Logical. TRUE = 17539, FALSE = 1853, NAs = 0.}


\item{constitutional.event}{Constitutional event (e.g., new constitution,
amendment, etc.). NA if no event took place during the year. Character with 8
distinct values. Most common: amendment (2231), interim (78), new (777). NAs =
16243.}


\item{has_mention}{Does the constitution mention democracy? (FALSE if no
constitution). Logical. TRUE = 9463, FALSE = 9929, NAs = 0.}


\item{mentions}{First 500 characters of the passages mentioning democracy.
(More than one mention may be concatenated together) Character with 1218
distinct values. Most common: and except so far as that provision o... (115),
on a democratic (63), Part I Chapter I General Declarat... (70). NAs = 9929.}


\item{GWc}{Gleditsch-Ward alphabetic country code. See Gleditsch and Ward
(1999). Character with 213 distinct values. Most common: CAN (251), UKG (799),
USA (233). NAs = 0.}


\item{cown}{Correlates of War numeric country code. Differs from GWn for a few
country-years. See Gleditsch and Ward (1999). Numeric. Max = 990, min = 2,
distinct = 213, mean = 389.543, sd = 259.79, NAs = 0.}


\item{polity_ccode}{Country code in Polity datasets. Differs from GWn for a few
country-years. See Gleditsch and Ward (1999). Numeric. Max = 990, min = 2,
distinct = 221, mean = 389.533, sd = 259.782, NAs = 0.}


\item{region}{Region. Character with 23 distinct values. Most common: Northern
Europe (1870), South America (2053), Western Europe (2210). NAs = 0.}


\item{continent}{Continent. Character with 5 distinct values. Most common:
Americas (4752), Asia (3937), Europe (6781). NAs = 0.}


\item{GW_startdate}{Date at which the state entered the system of states
according to Gleditsch and Ward, or NA if it has never been a member. Date. Max
= 2011-07-09, min = 1816-01-01, distinct = 166, NAs = 0.}


\item{GW_enddate}{Date at which the state ceased to be a member of the system
of states according to Gleditsch and Ward, or NA if it still exists. Date. Max
= 2006-06-04, min = 1830-09-22, distinct = 19, NAs = 17891.}


\item{microstate}{Indicator of whether the state is a microstate, according to
Gleditsch's list of microstates. Logical. TRUE = 1326, FALSE = 18066, NAs = 0.}


\item{lat}{Latitude. Numeric. Max = 64.963, min = -40.901, distinct = 210, mean
= 23.699, sd = 26.004, NAs = 0.}


\item{lon}{Longitude. Numeric. Max = 178.68, min = -175.198, distinct = 210,
mean = 5.571, sd = 67.087, NAs = 0.}


\item{in_cow}{Whether the country-year is in the Correlates of War system of
states. Logical. TRUE = 15480, FALSE = 3912, NAs = 0.}


\item{in_system}{Whether the country-year is in the Gleditsch-Ward system of
states. See Gleditsch and Ward (1999). Logical. TRUE = 17509, FALSE = 1883, NAs
= 0.}


}
}
\references{
Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate System Membership:
A Revised List of the Independent States since 1816." International
Interactions 25: 393-413. The list can be found at
\url{http://privatewww.essex.ac.uk/~ksg/statelist.html}
}
\seealso{
Other institutions: \code{\link{PIPE}}, \code{\link{lied}},
  \code{\link{svolik_institutions}}
}
\keyword{datasets}

