% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personal_scores_doc.R
\docType{data}
\name{personal_scores}
\alias{personal_scores}
\title{Latent variable index of personalism, v. 1.0}
\format{An object of class \code{data.frame} with 22597 rows and 8 columns.}
\usage{
personal_scores
}
\description{
A latent index of personal power, constructed from a variety of indicators of
personalism. See \\code{vignette("An_index_of_personal_power")} for more
information on how the index of personalism is constructed.
}
\section{Variable descriptions}{


\describe{ \item{country_name}{Standardized country name. This is the same
across all datasets in this package, so you can always join them by
country_name and year. Character with 215 distinct values. Most common:
Afghanistan (216), Denmark (216), Iran (Persia) (216), Morocco (216), Nepal
(216), Oman (275), Russia (Soviet Union) (216), Thailand (216), Turkey (Ottoman
Empire) (216), United Kingdom (216), United States of America (216). NAs = 0.}


\item{GWn}{Gleditsch-Ward numeric country code. See Gleditsch and Ward (1999).
Numeric. Max = 990, min = 2, distinct = 215, mean = 436.329, sd = 254.995, NAs
= 0.}


\item{year}{Year. Numeric. Max = 2015, min = 1741, distinct = 275, mean =
1940.34, sd = 53.959, NAs = 0.}


\item{in_system}{Whether the country-year is in the Gleditsch-Ward system of
states. See Gleditsch and Ward (1999). Logical. TRUE = 17177, FALSE = 5420, NAs
= 0.}


\item{z1}{Estimated index. Lower values indicate more personalism, so I usually
invert it. Numeric. Max = 2.514, min = -2.461, distinct = 4368, mean = -0.004,
sd = 0.891, NAs = 0.}


\item{se.z1}{Standard error of the estimate. Numeric. Max = 0.954, min = 0.241,
distinct = 4368, mean = 0.42, sd = 0.17, NAs = 0.}


\item{pct975}{Upper bound of the 95\% confidence interval. Numeric. Max =
3.467, min = -1.497, distinct = 4368, mean = 0.819, sd = 0.929, NAs = 0.}


\item{pct025}{Lower bound of the 95\% confidence interval. Numeric. Max =
1.561, min = -3.424, distinct = 4368, mean = -0.828, sd = 0.974, NAs = 0.}


}

function (x, ...)  UseMethod("end")
}
\keyword{datasets}

