% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world_doc.R
\docType{data}
\name{world}
\alias{world}
\title{World map data, in country-year format, for use with ggplot2.}
\format{An object of class \code{data.frame} with 12374 rows and 12 columns.}
\usage{
world
}
\description{
World map data, in country-year format, for use with ggplot2. Taken from
rworldmap and joined with columns for the standardized country name and the
Gleditsch-Ward country code.
}
\section{Variable descriptions}{


\describe{ \item{long}{Longitude. Numeric. Max = 180, min = -180, distinct =
10032, mean = 8.867, sd = 86.144, NAs = 0.}


\item{lat}{Latitude. Numeric. Max = 83.645, min = -90, distinct = 9907, mean =
16.708, sd = 36.429, NAs = 0.}


\item{order}{Plotting order. Numeric. Max = 795, min = 1, distinct = 795, mean
= 98.253, sd = 150.885, NAs = 0.}


\item{hole}{Whether the polygon describes a hole. Logical. TRUE = 12, FALSE =
12362, NAs = 0.}


\item{piece}{piece Factor with 30 levels. Most common: 1 (9732), 2 (917), 3
(427), 4 (232), 5 (179), 6 (116), 8 (112). NAs = 0.}


\item{id}{id Character with 242 distinct values. Most common: Antarctica (661),
Canada (795), Russia (Soviet Union) (598). NAs = 0.}


\item{group}{group Factor with 454 levels. Most common: Antarctica.1 (556),
Australia.1 (224), Brazil.1 (203), Canada.1 (272), China.1 (230), Russia.1
(447), United States of America.1 (233). NAs = 0.}


\item{year}{Year. Numeric. Max = 2012, min = 2012, distinct = 1, mean = 2012,
sd = 0, NAs = 0.}


\item{country_name}{Standardized country name. This is the same across all
datasets in this package, so you can always join them by country_name and year.
Character with 197 distinct values. Most common: Canada (795), Russia (Soviet
Union) (598), United States of America (444). NAs = 1786.}


\item{GWn}{Gleditsch-Ward numeric country code. See Gleditsch and Ward (1999).
Numeric. Max = 990, min = 2, distinct = 197, mean = 443.972, sd = 294.599, NAs
= 1786.}


\item{reg}{reg Character with 23 distinct values. Most common: Eastern Europe
(983), Northern America (1239), South America (919). NAs = 1786.}


\item{continent}{Continent. Character with 6 distinct values. Most common:
Africa (2370), Americas (2858), Asia (2395). NAs = 1786.}


}

function (x, ...)  UseMethod("end")
}
\keyword{datasets}

